% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_processing_utilities.R
\name{add.margins}
\alias{add.margins}
\title{Function used to add whitespace to image}
\usage{
add.margins(rst)
}
\arguments{
\item{rst}{Raster image}
}
\description{
This function adds whitespace to a raster image and forces the height/width
to be equal to the image diagonal. This way you can ensure that no part of the tissue
is cropped after a rotation transformation
}
