% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/image_processing.R
\name{AlignImages}
\alias{AlignImages}
\alias{AlignImages.Staffli}
\alias{AlignImages.Seurat}
\title{Automatic alignment of HE stained tissue images}
\usage{
AlignImages(
  object,
  indices = NULL,
  reference.index = NULL,
  use.masked = FALSE,
  verbose = FALSE
)

\method{AlignImages}{Staffli}(
  object,
  indices = NULL,
  reference.index = NULL,
  use.masked = FALSE,
  verbose = FALSE
)

\method{AlignImages}{Seurat}(
  object,
  indices = NULL,
  reference.index = NULL,
  use.masked = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{object}{A Seurat object}

\item{indices}{Integer: sample indices of images to align with reference}

\item{reference.index}{Integer: sample index of referece image}

\item{use.masked}{Setting this to TRUE will ignore any "processed" images available in the object and
apply the alignment from scratch.}

\item{verbose}{Print messages}
}
\value{
A Staffli object

A Seurat object
}
\description{
Image alignment or image registration consists in finding a rigid tranformation
function that remaps pixels between two images so that two images are aligned.
AlignImages allows you to align all images to a reference (any image present in the Seurat object)
which simplifies the interpretation of spatial heatmaps and can be useful for creating 3D
models. The transformation function is learned using the ICP (Iterative Closest Point) on two point
sets which defines the edges of the tissues in the HE images. Note that this alignment works best
for tissue sections that are intact, i.e. not cropped or folded. Also, because the method only used
the edges of the tissue for alignment, you might end up with strange results if the tissue
shape is symmetrical.
}
\examples{
\dontrun{
# Create a new Staffli object, mask, align and plot images
st.obj <- CreateStaffliObject(imgs, meta.data)
st.obj <- LoadImages(st.obj, verbose = TRUE) \%>\% MaskImages() \%>\% AlignImages()
plot(st.obj)
}
\dontrun{
# Load, mask, align and plot images
se <- LoadImages(se, verbose = TRUE) \%>\% MaskImages() \%>\% AlignImages()
ImagePlot(se)
}
}
