% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rigid_transformation.R,
%   R/image_processing_utilities.R
\name{apply.transform}
\alias{apply.transform}
\title{Apply rigid transformation to a set of points}
\usage{
apply.transform(map, set)

apply.transform(map, set)
}
\arguments{
\item{map}{List containing transformation matrices}

\item{set}{Matrix of x, y coordinates to be transformed}

\item{icp}{List containing transformation matrices}
}
\value{
Matrix of transformed x, y coordinates

Matrix of transformed x, y coordinates
}
\description{
Takes a list obtained with \code{\link{find.optimal.transform}} and
a matrix of x, y coordinates and returns the transformed x, y coordinates

Takes a list of obtained with \code{\link{find.optimal.transform}} and
a matrix of x, y coordinates and returns the transformed x, y coordinates
}
