% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_genes.R
\name{BinSpatialGenes}
\alias{BinSpatialGenes}
\title{Find genes with spatial structure in multiple datasets}
\usage{
BinSpatialGenes(object, spots = NULL, bin_method = c("kmeans", "rank"),
  slot = "data", features.include = NULL, nstart = 3,
  iter_max = 10, percentage_rank = 10, do_fisher_test = F,
  community_expectation = 5, verbose = FALSE)
}
\arguments{
\item{object}{Seurat object}

\item{spots}{Character vector with spot IDs to plot [default: all spots]}

\item{bin_method}{Chose between methods 'kmeans' or 'rank' to rank gene expression on}

\item{slot}{Which slot to pull the data from? [default: 'data']}

\item{features.include}{Features to include in the detection computation. By default,
most variable features are selected.}

\item{nstart}{kmeans: nstart parameter}

\item{iter_max}{kmeans: iter.max parameter}

\item{percentage_rank}{Percentage of top spots for binarization}

\item{do_fisher_test}{Perform Fisher test}

\item{community_expectation}{Degree expectation in spatial communities}

\item{verbose}{Print messages}
}
\value{
List of data.tables with results (see details)

list of data.frames with spatial scores for each gene
}
\description{
Rapid computation of genes that are spatially clustered in multiple samples.
}
\details{
Two ways are provided to identify spatial genes based on gene expression binarization.
Both methods are identicial except for how binarization is performed.
\itemize{
  \item{1. binarize: }{Each gene is binarized (0 or 1) in each cspot with \bold{kmeans} (k = 2) or based on \bold{rank} percentile}
  \item{2. network: }{Alll spots are connected through a k-nearest neighbor network}
  \item{3. contingency table: }{A contingency table is calculated based on all pairwise spot-spot interactions (0-0, 0-1, 1-0 or 1-1)}
  \item{4. For each gene an odds-ratio (OR) and fisher.test (optional) is calculated}
}
Additionally 2 other statistics are provided:
\itemize{
  \item{Number of spots with high expression (binary = 1)}
  \item{total and ratio of highly connected spots: }{ Spots with a connectivity higher than community_expectation}
}
By selecting a subset of likely spatial genes (e.g. highly variable genes) the function will be much faster.
}
\section{Acknowledgement}{

The function is based on the `binGetSpatialGenes` function from the `Giotto` R package,
but have been modified to work with `Seurat` objects and with multiple samples.
}

\examples{
spatgenes <- BinSpatialGenes(se)

}
