% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{ColorBlender}
\alias{ColorBlender}
\title{Squeeze 2 or 3 column feature data into the unit cube and converts into RGB space}
\usage{
ColorBlender(data, channels.use = NULL)
}
\arguments{
\item{data}{data.frame containing feature values and coordinates}

\item{channels.use}{Select channels to use for blending. Default is red, green and blue but the order can be shuffled.
For 2 features, the default is red and green. (options: "red", "green" and "blue")}
}
\description{
Squeeze 2 or 3 column feature data into the unit cube and converts into RGB space
}
