% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ConvertGeneNames}
\alias{ConvertGeneNames}
\title{Convert gene names}
\usage{
ConvertGeneNames(exprMat, annotation, id.column, replace.column)
}
\arguments{
\item{exprMat}{expression matrix}

\item{annotation}{data.frame with one id.column containing the current gene ids and a matching replace
column containing the new gene ids}

\item{id.column}{character vector of unique gene ids of the same type as in the Seurat object}

\item{replace.column}{character vector of unique gene ids to replace the old ids}
}
\value{
A Seurat object with covnerted gene ids
}
\description{
Convert gene names
}
