% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSVencoding.R
\name{create.cols.from.array}
\alias{create.cols.from.array}
\title{Creates HSV colors from an array}
\usage{
create.cols.from.array(data, d, features, cols, split.hsv, dark.theme,
  add.alpha)
}
\arguments{
\item{data}{data.frame with feature values}

\item{d}{array created with \code{create.array.from.vals} function}

\item{features}{Feature names}

\item{cols}{Custom colors}

\item{split.hsv}{Should the features be plotted separately?}

\item{dark.theme}{Used to select what channel the feature values should be encoded in}

\item{add.alpha}{Adds opacity to the output colors, defined by the scaled feature values}
}
\description{
If split.hsv = FALSE, the feature with the highest value in a spot will define the
color for that spot. The intensity of the color will depend on if dark.theme is active and
the magnitude of the feature value in that spot.
}
