% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_annotation.R
\name{Create_annotation}
\alias{Create_annotation}
\title{Create an annotation}
\usage{
Create_annotation(object, type, sampleNr, res)
}
\arguments{
\item{object}{Seurat object}

\item{type}{Image type, e.g. "raw" or "masked"}

\item{sampleNr}{Selected section/sample number}

\item{res}{Image resolution}
}
\description{
Create an annotation
}
