% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_grid_SC.R
\docType{data}
\name{CustomFacetWrap}
\alias{CustomFacetWrap}
\title{Custom ggproto object for facet wrap}
\format{
An object of class \code{CustomFacetWrap} (inherits from \code{FacetWrap}, \code{Facet}, \code{ggproto}, \code{gg}) of length 2.
}
\usage{
CustomFacetWrap
}
\description{
Custom ggproto object for facet wrap
}
\keyword{datasets}
