% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mask_functions.R
\name{EdgeMask}
\alias{EdgeMask}
\title{Masking function t}
\usage{
EdgeMask(im)
}
\arguments{
\item{im}{An image of class \code{cimg}}
}
\description{
This function takes an image of class \code{cimg} as input and returns
an image of class \code{pixset}. The method is typically suitable for images
with little background debris. Not suitable for 'Visium' images with fiducials.
}
