% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{expr.merger}
\alias{expr.merger}
\title{Merge expression matrices}
\usage{
expr.merger(exp.list, delim = "_", labels = NULL)
}
\arguments{
\item{exp.list}{list of expression matrices}

\item{delim}{delimiter used to separate coordinates in expression matrix headers}

\item{labels}{labels to use as suffix for the headers of each expression matrix}
}
\description{
This function takes a list of expression matrices as input with ST spots in columns
and genes in rows. All expression matrices will be merged into one matrix where all genes that are
present in any of the expression matrices will be included. If a there are no counts available for
a gene in one or more of the expression matrices, the counts for that gene will be set to 0 in those
expression matrices. By default, a unique number will be prefixed to the headers of each expression
matrix.
}
