% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_grid_SC.R
\name{facet_wrap_custom}
\alias{facet_wrap_custom}
\title{Custom facet wrap function}
\usage{
facet_wrap_custom(..., scale_overrides = NULL)
}
\arguments{
\item{...}{Parameters whcih can be passed to \code{facet_wrap}}

\item{scale_overrides}{List of "scale_override" objects}
}
\description{
Custom facet wrap function
}
