% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunNMF.R
\name{FactorGeneLoadingPlot}
\alias{FactorGeneLoadingPlot}
\title{Plot a bar chart with gene loadings}
\usage{
FactorGeneLoadingPlot(object, factor = 1, topn = 20,
  dark.theme = FALSE)
}
\arguments{
\item{object}{Seurat object}

\item{factor}{Factor to plot gene loadings for [default: 1]}

\item{topn}{Top genes to show [default: 20]}

\item{dark.theme}{Use a dark theme for the plot}
}
\description{
This function takes a `Seurat`object as input with a 'NMF' reduction slot
and plots a bar chart of gene weights (loadings) for the top most driving genes.
}
