% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3D_viusalization.R
\name{FeaturePlot3D}
\alias{FeaturePlot3D}
\title{Plots the values of a feature in 3D}
\usage{
FeaturePlot3D(object, features, slot = "data", pt.size = 0.8,
  cols = NULL, add.margins = 2)
}
\arguments{
\item{object}{Seurat object}

\item{slot}{Which slot to pull the data from? [default: 'data']}

\item{feature}{Feature to plot}
}
\description{
Plots the values of a feature in 3D
}
