% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{FilterObjects}
\alias{FilterObjects}
\title{Get the names of objects within a Seurat object that are of a certain class}
\usage{
FilterObjects(object, classes.keep = c("Assay", "DimReduc"))
}
\arguments{
\item{object}{A Seurat object}

\item{classes.keep}{A vector of names of classes to get}
}
\value{
A vector with the names of objects within the Seurat object that are of class \code{classes.keep}
}
\description{
Get the names of objects within a Seurat object that are of a certain class
}
