% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_genes.R
\name{fish_function2}
\alias{fish_function2}
\title{fish_function2}
\usage{
fish_function2(A, B, C, D)
}
\arguments{
\item{A, B, C, D}{Vectors used for Fishers test}
}
\description{
Perform fisher exact test
}
