% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_processing_utilities.R
\name{generate.map.affine}
\alias{generate.map.affine}
\title{Generate a map function}
\usage{
generate.map.affine(tr, forward = FALSE)
}
\arguments{
\item{tr}{results obtained with \code{\link{find.optimal.transform}}}

\item{forward}{Should the forward transformation be computed?}
}
\value{
A transformation function that takes x and y coordinates as input and outputs a
list of warped x, y coordinates
}
\description{
Given two sets of points (2D) this function will create a rigid transformation
function that minimized the rmse of the two points sets. Because we use the backward
transformation function for image warping, we need to compute the forward transformation
function for pixel coordinate warping. If set1 is (x, y) and set2 is the target (x', y')
point set, we want to find the backward transformation function M(x', y') -> (x, y).
The forward transformation will be the inverse of M, i.e. M^-1(x, y) -> (x', y').
The iterative closest point algorithm used to find the best rigid transformation function
may not find the best solution if the image is reflected. For this reason we calculate
the transofmration function for all different combinations of reflections and select the
function with the lowest rmse between the aligned set and reference set.
}
