% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_processing_utilities.R
\name{generate.map.rot}
\alias{generate.map.rot}
\title{Create transformation function}
\usage{
generate.map.rot(tr, forward = FALSE)
}
\arguments{
\item{tr}{Forward transformation matrix}

\item{forward}{Logical: sets algorithm to 'forwar', otherwise 'backward'}
}
\description{
Creates a function that takes x,y values as input and applies a transformation
}
