% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get.edges}
\alias{get.edges}
\title{Obtain edges of binaryh mask stores in masked.masks list}
\usage{
get.edges(object, index = 1, verbose = FALSE, type = "masked.masks")
}
\arguments{
\item{object}{Seurat object}

\item{index}{Sample index}

\item{verbose}{Print messages}

\item{Input}{image type to run edge detection on [default: 'masked.masks']}
}
\description{
Obtain edges of binaryh mask stores in masked.masks list
}
