% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_processing_utilities.R
\name{get.edges.Seurat}
\alias{get.edges.Seurat}
\title{Finds edges of a mask and returns a set of x,y coordinates
aligned along the edges}
\usage{
\method{get.edges}{Seurat}(object, index = 1, verbose = FALSE,
  type = "masked.masks")
}
\arguments{
\item{object}{Seurat object}

\item{index}{Sample index}

\item{verbose}{Print messages}
}
\value{
A list of data.frames with edge coordinates
}
\description{
Finds edges of a mask and returns a set of x,y coordinates
aligned along the edges
}
\examples{
# Mask, align and plot edges for image 2 from a Seurat object
edges <- LoadImages(se, verbose = TRUE) \%>\% MaskImages() \%>\% get.edges(index = 2)
plot(as.raster(edges) \%>\% as.cimg())
}
