% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{GetCoords}
\alias{GetCoords}
\title{Extract table of x, y coordinates from ST count matrix headers}
\usage{
GetCoords(spotnames, delim = "x")
}
\arguments{
\item{spotnames}{x, y coordinates separated by delimiter (default "x")}

\item{delim}{delimiter}
}
\value{
data.frame with x, y coordinates and optionally a sample column
}
\description{
Extract table of x, y coordinates from ST count matrix headers
}
