% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_utilities.R
\name{GetCropWindows}
\alias{GetCropWindows}
\title{Get crop geometries}
\usage{
GetCropWindows(
  object,
  group.by = NULL,
  groups.to.keep = NULL,
  keep.all.spots = FALSE,
  xy_padding = 50
)
}
\arguments{
\item{object}{A Seurat object created with STutility}

\item{group.by}{A meta.data column to group the spots by. Only character 
vector or factors are allowed. Selects "labels" by default.}

\item{groups.to.keep}{Provide a caharcter vector of groups to keep.}

\item{keep.all.spots}{Keep all spots within selected crop windows and not just the spots 
belonging to the predefined groups. All spots that are present in multiple crop windows will be removed.}

\item{xy_padding}{Increase the crop area in all directions. Given in pixels.}
}
\value{
A list of "crop geometries" to be used with \code{\link{CropImages}})
}
\description{
This function can be used to define "crop windows" based on sets 
of target spots.
}
