% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_genes.R
\name{GetSpatNet}
\alias{GetSpatNet}
\title{Create Spatial Networks}
\usage{
GetSpatNet(object, nNeighbours = NULL, maxdist = NULL, minK = 0)
}
\arguments{
\item{object}{Seurat object}

\item{nNeighbours}{Number of nearest neighbours to calculate for each spot. The default
number of neighbours is 6 for the 'Visium' platform and 4 for the '1k' and '2k' platforms.}

\item{maxdist}{Distance cut-off for nearest neighbours to consider. The default is 1.5 for the
'Visium' and '2k' platforms and 2 for the '1k' platform.}

\item{minK}{Minimum nearest neigbhours if maxdist is not provided [default: 0]}
}
\description{
Create spatial networks based on spot center to center distances from a Seurat object.
}
\examples{
\dontrun{
spatial.networks <- GetSpatNet(se)
}
}
