% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Staffli.R
\name{GetStaffli,Seurat-method}
\alias{GetStaffli,Seurat-method}
\title{Method used to extract a Staffli object from the tools slot of a
Seurat object}
\usage{
\S4method{GetStaffli}{Seurat}(object)
}
\arguments{
\item{object}{A Seurat object}
}
\description{
Method used to extract a Staffli object from the tools slot of a
Seurat object
}
