% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D_point_pattern.R
\name{grid.from.staffli}
\alias{grid.from.staffli}
\title{Creates 2D point patterns for a set of images}
\usage{
grid.from.staffli(
  object,
  type,
  limit = 0.3,
  maxnum = 1000,
  edges = FALSE,
  custom.edge.detector = NULL
)
}
\arguments{
\item{object}{Staffli object}

\item{type}{Sets the image type to run segmentation on}

\item{limit}{Sets the intensity threshold in the interval [0, 1]}

\item{maxnum}{Integer value specifying the maximum number of points}

\item{edges}{Extracts the coordinates of the edges instead}

\item{custom.edge.detector}{Custom function used to detect edges}
}
\description{
Creates 2D point patterns for a set of images
}
