% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSVencoding.R
\name{hsv2hex}
\alias{hsv2hex}
\title{Tranforms hsv color codes into hex color code}
\usage{
hsv2hex(col)
}
\arguments{
\item{col}{vector of length 3 containing values for "h", "s" and "v"}
}
\description{
Tranforms hsv color codes into hex color code
}
