% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSVencoding.R
\name{HSVOverlay}
\alias{HSVOverlay}
\title{Overlay HSV encoded features on HE images}
\usage{
HSVOverlay(object, features, sampleids = 1, rescale = TRUE,
  spots = NULL, ncols.features = NULL, ncols.samples = NULL,
  type = NULL, min.cutoff = NA, max.cutoff = NA, slot = "data",
  pt.size = 2, pt.alpha = 1, add.alpha = FALSE, shape.by = NULL,
  palette = NULL, cols = NULL, split.hsv = FALSE,
  dark.theme = FALSE, sample.label = TRUE, show.sb = TRUE,
  custom.theme = NULL, verbose = FALSE, ...)
}
\arguments{
\item{object}{Seurat object}

\item{features}{\itemize{
    \item An \code{Assay} feature (e.g. a gene name - "MS4A1")
    \item A column name from meta.data (e.g. mitochondrial percentage - "percent.mito")
}}

\item{sampleids}{Names of samples to plot}

\item{rescale}{Rescale each feature column separately from 0 to 1 range. If set to FALSE, all feature columns
will be scaled together from 0 to 1 and preserve the relative differencies}

\item{spots}{Character vector with spot IDs to plot [default: all spots]}

\item{ncols.features}{Number of columns passed to \code{\link{FeatureOverlay}}. For example,
if you are plotting 4 features, `ncols.features = 2` will arrange the \code{\link{FeatureOverlay}}
plots into a 2x2 grid [default: `length(features)`]. (see \emph{Arrange plots*} for a detailed description)}

\item{ncols.samples}{Number of columns in the layout grid for the samples. For example,
if you are plotting 4 samples, `ncols.samples = 2` will arrange the plots obtained
from \code{\link{FeatureOverlay}} plots into a 2x2 grid [default: `1`].
(see \emph{Arrange plots*} for a detailed description)}

\item{type}{Image type to plot on. Here you can specify any of the images available in your Seurat object. To get this list you can
run the \code{\link{rasterlists}} function on your Seurat object. If the type is not specified, the images will be prioritized in the following
order if they are available; "processed", "masked" and "raw".}

\item{min.cutoff}{Vector of minimum and maximum cutoff values for each feature, may specify quantile in the form of 'q##' where '##'
is the quantile (eg, 'q1', 'q10'). This can be useful if you have outlier values that skew the colorscale in the plot. For example, if you specify
'q1', you will trim of values below the 1st percentile. [default: no cuttoffs]}

\item{max.cutoff}{Vector of minimum and maximum cutoff values for each feature, may specify quantile in the form of 'q##' where '##'
is the quantile (eg, 'q1', 'q10'). This can be useful if you have outlier values that skew the colorscale in the plot. For example, if you specify
'q1', you will trim of values below the 1st percentile. [default: no cuttoffs]}

\item{slot}{Which slot to pull expression data from? [dafault: 'data']}

\item{pt.size}{Point size of each ST spot [default: 1]}

\item{pt.alpha}{Opacity of each ST spot [default: 1]}

\item{add.alpha}{Adds opacity to spots scaled by feature values. This will disable the pt.alpha parameter}

\item{shape.by}{String specifying the column where the shaping label is stored}

\item{palette}{Color palette used for spatial heatmap (see \code{palette.select(info = T)} for available options).
Disabled if a color vector is provided (see \code{cols} below).}

\item{cols}{A vector of colors to use for colorscale, e.g. \code{cols = c("blue", "white", "red")} will
create a gradient color scale going from blue to white to red. This options will deactivate the \code{palette}
option.}

\item{split.hsv}{Should the HSV colored features be split into separate plots? [default: FALSE]}

\item{dark.theme}{Switches color of scalebar to 'white'}

\item{sample.label}{Should the sample label be included in the image? [default: TRUE]}

\item{show.sb}{Should a scalebar be drawn? [default: TRUE]}

\item{custom.theme}{Object of class 'theme' used to change the background theme (see for example \url{https://ggplot2.tidyverse.org/reference/theme.html})}

\item{verbose}{Print messages}

\item{...}{Parameters passed to DimOverlay}
}
\description{
Graphs the selected features as a HSVplot on a 2D grid of spots overlaid on top of an HE images.
Only numerical features are accepted, e.g. genes or dimensionality reduction output vectors. If you
want to draw dimentionality reduction vectors you need to specify the whole names of the vectors, e.g.
`features = c("factor_1", "factor_2")` for the two first NMF factors.
}
\details{
NOTE that this function draws sample 1 as default, but can take multiple samples as well using the `sampleids argument`.

It is typically difficult to explore details in the HE image when diplaying multiple samples side by side,
so we recommend to draw the plots for one sample at the time. If you have higher resolution images,
it could also take significant time to draw the plots.
}
\section{Arrange plots}{


The `ncols.features` argument will determine how each subplot called using
\code{\link{DimOverlay}} is arranged and will by default put all dims in 1 row, i.e.
`ncols.features = length(features)`. The `ncols.samples` argument will determine how these subplots
are arranged and will by default use 1 column, meaning that each subplot is put in its own row.
The output layout matrix would then have the dimensions `length(samples)xlength(features)`



The `ncols.features` argument will determine how each subplot is arranged and will by default put all features in 1 row, i.e.
`ncols.features = length(features)`. The `ncols.samples` argument will determine how these subplots
are arranged and will by default use 1 column, meaning that each subplot is put in its own row.
The output layout matrix would then have the dimensions `length(samples)xlength(features)`
}

\section{Splitting categorical features}{

If you are plotting a categorical feature, e.g.cluster labels, you have the option to split each label into facets using \code{split.labels=TRUE}.
This is very useful if you have many different labels which can make it difficult to distinguish the different colors.
}

\examples{
# Load images
se <- se \%>\% SCTransfrom() \%>\% LoadImages() \%>\% RunNMF()

# Overlay first two NMF factors on the first two tissue sections
HSVPlot(se, features = c("factor_1", "factor_2"), sampleids = 1:2)

}
