% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3D_viusalization.R
\name{HSVPlot3D}
\alias{HSVPlot3D}
\title{Plots the values of a feature in 3D}
\usage{
HSVPlot3D(object, features, mode = c("cloud", "spots"), zcoords = NULL,
  spots = NULL, min.cutoff = NA, max.cutoff = NA, slot = "data",
  blend = FALSE, pt.size = NULL, pt.alpha = 1, cols = NULL,
  add.alpha = FALSE, add.margins = 0, channels.use = NULL,
  scene = "scene1", return.data = FALSE, dark.theme = FALSE,
  verbose = FALSE)
}
\arguments{
\item{object}{Seurat object}

\item{features}{\itemize{
    \item An \code{Assay} feature (e.g. a gene name - "MS4A1")
    \item A column name from meta.data (e.g. mitochondrial percentage - "percent.mito")
    \item A column name from dimensionality reduction output (e.g. principal component 1 - "PC_1")
}}

\item{zcoords}{Vector of z coordinates with the same length as the number of sections in the dataset [default: 1:#sections]}

\item{spots}{Vector of spots to plot (default is all spots)}

\item{min.cutoff, max.cutoff}{Vector of minimum and maximum cutoff values for each feature, may specify quantile in the form of 'q##' where '##'
is the quantile (eg, 'q1', 'q10'). This can be useful if you have outlier values that skew the colorscale in the plot. For example, if you specify
'q1', you will trim of values below the 1st percentile. [default: no cuttoffs]}

\item{slot}{Which slot to pull the data from? [default: 'data']}

\item{blend}{Scale and blend expression values to visualize coexpression of two features (this options will override other coloring parameters).
See 'Blending values' below for a more thourough description.}

\item{pt.size}{Sets the size of points in the 3D plot}

\item{pt.alpha}{Sets the opacity of the points}

\item{cols}{Character vector of colors with equal length to the number of features. These colors will override
the selection of HSV colors and can therefore not be encoded in the same way. Instead of tuning the saturation/value
parameters we can add an alpha channel, making spots with values close to zero completely transparent. This has to
be activated by setting `add.alpha = TRUE`.}

\item{add.alpha}{Adds opacity to colors.}

\item{add.margins}{Add margins along z axis to push sections closer to each other}

\item{channels.use}{Color channels to use for blending. Has to be a character vector of length 2 or 3 with "red", "green" and "blue"
color names specified [default: c("red", "green", "blue)]}

\item{scene}{Give the scene a name to allow for multiple subplots}

\item{return.data}{return the data.frame with x,y coordinates and interpolated values
instead of plotting}

\item{dark.theme}{Draws the plot with a dark theme}

\item{verbose}{Print messages}
}
\description{
Plots the values of a feature in 3D
}
