% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunNMF.R
\name{ica_init}
\alias{ica_init}
\title{Initiate NMF using ICA}
\usage{
ica_init(A, k, ica.fast = F)
}
\arguments{
\item{A}{input matrix}

\item{k}{number of components to compute}

\item{ica.fast}{Should a fast implementation of ICA be used?}
}
\description{
Initiate NMF using ICA
}
