% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_processing.R
\name{ImagePlot}
\alias{ImagePlot}
\title{Function used to plot HE images obtained with \code{\link{LoadImages}}}
\usage{
ImagePlot(object, indices = NULL, type = NULL, method = "viewer",
  ncols = NULL, annotate = TRUE, darken = FALSE)
}
\arguments{
\item{object}{Seurat object}

\item{indices}{Image indices to select}

\item{type}{Specify which image to display [options: "raw", "masked" or "processed"]}

\item{method}{Specify display method (raster or viewer).}

\item{ncols}{Number of columns in output grid of images}

\item{annotate}{Will put a unique id in the top left corner}

\item{darken}{Switches the background to black}
}
\description{
Function used to plot HE images obtained with \code{\link{LoadImages}}
}
