% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Staffli.R
\name{iminfo,Staffli-method}
\alias{iminfo,Staffli-method}
\title{Method for extracting the HE image dimensions from a 'Staffli' object
 @param object A Staffli object}
\usage{
\S4method{iminfo}{Staffli}(object)
}
\description{
Method for extracting the HE image dimensions from a 'Staffli' object
 @param object A Staffli object
}
