% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{iminfo}
\alias{iminfo}
\title{Extracts a list of image info from a Staffli object}
\usage{
iminfo(object)
}
\arguments{
\item{object}{A Staffli object}
}
\description{
Extracts a list of image info from a Staffli object
}
