% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_from_table.R
\name{InputFromTable}
\alias{InputFromTable}
\title{Create S4 clas Seurat object from info table.}
\usage{
InputFromTable(infotable, transpose = TRUE, topN = 0,
  min.gene.count = 0, min.gene.spots = 0, min.spot.feature.count = 0,
  min.spot.count = 0, annotation = NULL, id.column = NULL,
  replace.column = NULL, platform = "Visium", scaleVisium = NULL,
  verbose = TRUE, ...)
}
\arguments{
\item{infotable}{table with paths to count files and metadata. See details below for more information.}

\item{topN}{OPTIONAL: Filter out the top most expressed genes}

\item{min.gene.count}{filter away genes that has a total read count below this threshold}

\item{min.gene.spots}{filter away genes that is not expressed below this number of capture spots}

\item{min.spot.feature.count}{filter away capture spots that contains a total feature count below this threshold}

\item{min.spot.count}{filter away capture spots that contains a total read count below this threshold}

\item{annotation}{data.frame with one id.column containing the current gene ids and a matching replace
column containing the new gene ids}

\item{id.column}{character vector of unique gene ids of the same type as in the Seurat object}

\item{replace.column}{character vector of unique gene ids to replace the old ids}

\item{scaleVisium}{10X visium scale factor for pixel coordinates mtaching the "tissue_hires_image.png" files [required].
If a numeric value is given, it is assumed that all samples have the same scaling factor. Alternatively, an additional
column named "scaleVisium" can be provided with a scaling factor for each sample, or a column named "json" with paths to
the "scalefactors_json.json" files.}

\item{verbose}{Print messages}

\item{...}{parameters passed to \code{\link{CreateSeuratObject}}}
}
\description{
This function is a wrapper to create a complete S4 object with all the samples and metadata
}
\details{
This wrapper function has been written
}
