% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_genes.R
\name{kmeans_binarize}
\alias{kmeans_binarize}
\title{kmeans_binarize}
\usage{
kmeans_binarize(x, nstart = 3, iter.max = 10)
}
\arguments{
\item{x}{Expression vector}

\item{nstart}{Number of random sets}

\item{iter.max}{Maximum number of allowed iterations}
}
\description{
Create binarized scores using kmeans
}
