% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_utilities.R
\name{labelRGB}
\alias{labelRGB}
\title{Select spots by rgb values after ST.DimPlot}
\usage{
labelRGB(
  object,
  dimPlot,
  label1 = "label1",
  label2 = NULL,
  red1 = c(0, 255),
  green1 = c(0, 255),
  blue1 = c(0, 255),
  red2 = c(0, 255),
  green2 = c(0, 255),
  blue2 = c(0, 255)
)
}
\arguments{
\item{object}{A Seurat object}

\item{dimPlot}{A 'ggplot' object created with \code{ST.DimPlot}}

\item{label1, label2}{Optional names of the 1st and 2nd groups of capture-spots}
}
\description{
Select spots by rgb values after ST.DimPlot
}
