% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/image_processing.R
\name{LoadImages}
\alias{LoadImages}
\alias{LoadImages.Staffli}
\alias{LoadImages.Seurat}
\title{Function used to read HE images in jpeg or png format}
\usage{
LoadImages(
  object,
  image.paths = NULL,
  xdim = 400,
  crop.to.fiducials = FALSE,
  crop.scale.factors = c(9, 10, 10, 8),
  verbose = TRUE,
  time.resolve = FALSE
)

\method{LoadImages}{Staffli}(
  object,
  image.paths = NULL,
  xdim = 400,
  crop.to.fiducials = FALSE,
  crop.scale.factors = c(9, 10, 10, 8),
  verbose = TRUE,
  time.resolve = TRUE
)

\method{LoadImages}{Seurat}(
  object,
  image.paths = NULL,
  xdim = 400,
  crop.to.fiducials = FALSE,
  crop.scale.factors = c(9, 10, 10, 8),
  verbose = TRUE,
  time.resolve = TRUE
)
}
\arguments{
\item{object}{Seurat or Staffli object}

\item{image.paths}{Paths to HE images. This is only required if image paths are missing in the Seurat object.}

\item{xdim}{Sets the pixel width for scaling, e.g. 400 (maximum allowed width is 2000 pixels)}

\item{crop.to.fiducials}{Set to TRUE if you want to crop out background from the images outside the fiducials.}

\item{crop.scale.factors}{Numeric vector of length 4 providing a scaling factor for each side of the image.
The scaling factors define the number of spot widths away from the capture area to include in the cropped
image if  `crop.to.fiducials` is set to TRUE. The default values are c(9, 10, 10, 8) which corresponds to
left, top, right and bottom.}

\item{verbose}{Print messages}

\item{time.resolve}{Activate to stop R from loading raw images into memory}
}
\value{
A Staffli object

A Seurat object
}
\description{
Function used to read HE images in jpeg or png format
}
\examples{
\dontrun{
# Create a new Staffli object and plot images
st.obj <- CreateStaffliObject(imgs, meta.data)
st.obj <- LoadImages(st.obj, verbose = TRUE)
plot(st.obj)
}

\dontrun{
# Load images into a Seurat object and plot images
se <- LoadImages(se, verbose = TRUE)
ImagePlot(se)
}
}
