% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_annotation.R
\name{make.plot}
\alias{make.plot}
\title{Used for the manual annotation tool, returns plot and coordinate IDs for selected sample}
\usage{
make.plot(object, sampleNr, spotAlpha, Labels, SpotSize, res, ann)
}
\arguments{
\item{object}{Seurat object}

\item{sampleNr}{Sample to be plotted}

\item{spotAlpha}{geom_point opacity}

\item{Labels}{labels included in the input seurat object}

\item{SpotSize}{geom_point size}

\item{res}{resolution of the image}
}
\description{
Used for the manual annotation tool, returns plot and coordinate IDs for selected sample
}
\keyword{internal}
