% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/image_processing.R
\name{ManualAlignImages}
\alias{ManualAlignImages}
\alias{ManualAlignImages.Staffli}
\alias{ManualAlignImages.Seurat}
\title{Manual alignment of images}
\usage{
ManualAlignImages(object, type = NULL, reference.index = 1,
  edges = TRUE, verbose = FALSE, maxnum = 1000)

\method{ManualAlignImages}{Staffli}(object, type = "masked.masks",
  reference.index = 1, edges = TRUE, verbose = FALSE,
  maxnum = 1000)

\method{ManualAlignImages}{Seurat}(object, type = "masked.masks",
  reference.index = 1, edges = TRUE, verbose = FALSE,
  maxnum = 1000)
}
\arguments{
\item{object}{Seurat object}

\item{reference.index}{Specifies reference sample image for alignment(default: 1)}

\item{edges}{Uses the tissue edges as points set for alignment}

\item{verbose}{Print messages}

\item{maxnum}{Maximum grid number}

\item{Image}{type used for alignment}
}
\value{
A Staffli object

A Seurat object
}
\description{
Creates an interactive shiny application to align images manually
}
\examples{
# Create a new Staffli object, mask, align and plot images (will start an interactive shiny session)
st.obj <- CreateStaffliObject(imgs, meta.data)
st.obj <- LoadImages(st.obj, verbose = TRUE) \%>\% MaskImages() \%>\% ManualAlignImages()
plot(st.obj)
# Load, mask, align and plot images (will start an interactive shiny session)
se <- LoadImages(se, verbose = TRUE) \%>\% MaskImages() \%>\% ManualAlignImages()
ImagePlot(se)
}
