% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_annotation.R
\name{ManualAnnotation}
\alias{ManualAnnotation}
\title{Manual annotation tool via shiny
This function takes a seurat object with stored image locations and opens up the manual selection tool in the default browser}
\usage{
ManualAnnotation(object, res = 1500, verbose = FALSE)
}
\arguments{
\item{object}{Seurat object}
}
\description{
Manual annotation tool via shiny
This function takes a seurat object with stored image locations and opens up the manual selection tool in the default browser
}
