% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rigid_transformation.R
\name{mirror}
\alias{mirror}
\title{Creates a transformation matrix that mirrors an object
in 2D along either the x axis or y axis around its
center of mass}
\usage{
mirror(mirror.x = FALSE, mirror.y = FALSE, center.cur)
}
\arguments{
\item{mirror.x, mirror.y}{Logical specifying whether or not an
object should be reflected}

\item{center.cur}{Coordinates of the current center of mass}
}
\description{
Creates a transformation matrix that mirrors an object
in 2D along either the x axis or y axis around its
center of mass
}
