% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunNMF.R
\name{NMFPlot}
\alias{NMFPlot}
\title{Plot spatial heatmap of NMF factors}
\usage{
NMFPlot(object, factors = 1:5)
}
\arguments{
\item{object}{Seurat object}

\item{factors}{Specify factors to plot (integer vector)}
}
\description{
Plot spatial heatmap of NMF factors
}
