% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunNMF.R
\name{NMFRidgePlot}
\alias{NMFRidgePlot}
\title{Plot density distribution of factor values}
\usage{
NMFRidgePlot(object, factors = 1:5, top.n.features = 10)
}
\arguments{
\item{object}{Seurat object}

\item{factors}{Specify factors to display (default 1-5)}

\item{top.n.features}{Number of features to include in bar plot}
}
\description{
Plot density distribution of factor values
}
