% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_utilities.R
\name{obtain.array.coords}
\alias{obtain.array.coords}
\title{Obtain array coordinates}
\usage{
obtain.array.coords(st.object, data, image.type, spots)
}
\arguments{
\item{data}{Data.frame with array coordinates}

\item{image.type}{One of "raw", "masked" or "processed"}

\item{spots}{Spots to subset data on}

\item{object}{A Seurat object}
}
\description{
Obtain array coordinates
}
