% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_genes.R
\name{OR_function2}
\alias{OR_function2}
\title{OR_function2}
\usage{
OR_function2(A, B, C, D)
}
\arguments{
\item{A, B, C, D}{Vectors to used for calculation of odds ratio}
}
\description{
Calculate odds-ratio
}
