% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_from_table.R
\name{parse.spot.file}
\alias{parse.spot.file}
\title{Parse the output from the ST spot detector tool}
\usage{
parse.spot.file(path, delim = "\\t", ...)
}
\arguments{
\item{path}{Path to spot files}

\item{delim}{delimiter}
}
\description{
Parse the output from the ST spot detector tool
}
\keyword{internal}
