% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Staffli.R
\name{plot,Staffli,ANY-method}
\alias{plot,Staffli,ANY-method}
\title{Plot method for Staffli objects}
\usage{
\S4method{plot}{Staffli,ANY}(x, type = NULL, ...)
}
\arguments{
\item{x}{A Staffli object}

\item{type}{Image type to draw on}

\item{...}{Additional aprameters passed to points}
}
\description{
Plot method for Staffli objects
}
