% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Staffli.R
\name{plot,Staffli,missing-method}
\alias{plot,Staffli,missing-method}
\alias{plot,Staffli,Staffli-method}
\title{Plot method for Staffli objects}
\usage{
\S4method{plot}{Staffli,missing}(x, y = NULL, type = NULL, ...)
}
\arguments{
\item{x}{object to use for plotting.}

\item{y}{missing}

\item{type}{type of images to use as background for plot, e.g. "raw", "masked"}

\item{...}{additional parameters passed to \code{points}}
}
\description{
Plot method for Staffli objects
}
