% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_utilities.R
\name{prep.sb}
\alias{prep.sb}
\title{Prep scalebar}
\usage{
prep.sb(st.object, data, data.type, indices, split.labels, features, dims,
  show.sb)
}
\arguments{
\item{st.object}{An object of class 'Staffli'}

\item{data}{data.frame for plotting}

\item{data.type}{String specifying the class of the variable used for splitting}

\item{indices}{Integer vector specifying sample indices to include in the plot}

\item{split.labels}{Split plots}

\item{features}{Features included in data}

\item{dims}{List of image dimensions}

\item{show.sb}{Return NULL if FALSE}
}
\value{
A data.frame containing positions of a scale bar used for plotting
}
\description{
Prep scalebar
}
