% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STUtility.R, R/hello.R
\name{prepare.counts}
\alias{prepare.counts}
\title{Create S4 object}
\usage{
prepare.counts(paths, transpose = TRUE, topN = 0, th.gene = 10,
  th.spot = 10)

prepare.counts(paths, transpose = TRUE, topN = 0, th.gene = 10,
  th.spot = 10)
}
\arguments{
\item{paths}{file paths to count files.}

\item{transpose}{set TRUE if count files are in the form of genes as columns and spatial
coordiantes as rows.}

\item{topN}{OPTIONAL: Filter out the top most expressed genes}

\item{th.gene}{OPTIONAL: Filter genes with total counts across all samples below threshold}

\item{th.spot}{OPTIONAL: Filter `"capture-spots"` with total count values below threshold}

\item{sampleTable}{table with paths to count files and metadata}

\item{transpose}{set TRUE if count files are in the form of genes as columns and spatial
coordiantes as rows.}

\item{topN}{Filter out the top most expressed genes}

\item{th.gene}{Filter genes with total counts across all samples below threshold}

\item{th.spot}{Filter `"spots"` with total count values below threshold}
}
\description{
Create S4 object

Create summarized experiment
}
