% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{qc.corr}
\alias{qc.corr}
\title{Correlation plots between all samples}
\usage{
qc.corr(se, metric = "transcripts")
}
\arguments{
\item{se}{S4 object}
}
\description{
Correlation plots between all samples
}
