% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{qc.hist}
\alias{qc.hist}
\title{QC histograms on the number of unqiue transcripts and genes per spot}
\usage{
qc.hist(se, metric = "transcripts")
}
\arguments{
\item{se}{S4 object}
}
\description{
QC histograms on the number of unqiue transcripts and genes per spot
}
