% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{qc.scatter}
\alias{qc.scatter}
\title{function(x, y, ...) used to plot the contents of each panel of the display in pairs function}
\usage{
qc.scatter(x, y)
}
\arguments{
\item{x}{x coordinate}

\item{y}{y coordinate}
}
\value{
panel for pairs function
}
\description{
function(x, y, ...) used to plot the contents of each panel of the display in pairs function
}
\keyword{internal}
