% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{qcGenes}
\alias{qcGenes}
\title{Quality metrics for genes}
\usage{
qcGenes(object, delim = "x", labels = NULL)
}
\arguments{
\item{object}{expression matrix with spots in columns and genes in rows or Seurat object}

\item{delim}{delimiter}

\item{labels}{character vector with labels corresponding to dataset IDs in merged expression matrix [optional]}
}
\value{
data.frame with gene names and quality metrics
}
\description{
Quality metrics for genes
}
