% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{qcSamples}
\alias{qcSamples}
\title{Quality metrics for samples}
\usage{
qcSamples(exprMat, delim = "x", labels = NULL)
}
\arguments{
\item{exprMat}{expression matrix with spots in columns and genes in rows}

\item{delim}{delimiter}

\item{labels}{character vector with labels corresponding to datasets in merged expression matrix}
}
\value{
data.frame with quality metrics per sample
}
\description{
Quality metrics for samples
}
