% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{qcSpots}
\alias{qcSpots}
\title{Quality metrics for spots}
\usage{
qcSpots(exprMat, delim = "x")
}
\arguments{
\item{exprMat}{expression matrix with spots in columns and genes in rows}

\item{delim}{delimiter for expression matrix headers}
}
\value{
data.frame with x, y coordinates and quality metrics
}
\description{
Quality metrics for spots
}
