% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_genes.R
\name{rank_binarize}
\alias{rank_binarize}
\title{rank_binarize}
\usage{
rank_binarize(x, max_rank = 200)
}
\arguments{
\item{x}{Expression vector}

\item{max_rank}{Maximum gene rank to consider}
}
\description{
Create binarized scores using arbitrary rank of top genes
}
