% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3D_viusalization.R
\name{rasterize_scatter}
\alias{rasterize_scatter}
\title{Assign points to a grid}
\usage{
rasterize_scatter(scatter, r, nx)
}
\arguments{
\item{scatter}{data.frame with x,y coordinates for a point pattern}

\item{r}{A 'raster' object}

\item{nx}{Number of cells to interpolate over across the x axis}
}
\value{
data.frame object with point pattern and associated grid.cell number
}
\description{
takes a point pattern and a raster object as input and
assigns each point to a cell of the raster object.
}
