% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Staffli.R
\docType{methods}
\name{rasterlists,Seurat-method}
\alias{rasterlists,Seurat-method}
\title{Method to extract the available image types from a Seurat object}
\usage{
\S4method{rasterlists}{Seurat}(object)
}
\arguments{
\item{object}{A Seurat object}
}
\description{
Method to extract the available image types from a Seurat object
}
