% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Staffli.R
\docType{methods}
\name{rasterlists,Staffli-method}
\alias{rasterlists,Staffli-method}
\title{Method to extract the available image types from a Staffli object}
\usage{
\S4method{rasterlists}{Staffli}(object)
}
\arguments{
\item{object}{A Staffli object}
}
\description{
Method to extract the available image types from a Staffli object
}
