% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/Staffli.R
\docType{methods}
\name{rasterlists}
\alias{rasterlists}
\alias{rasterlists,Staffli-method}
\alias{rasterlists,Staffli,Staffli-method}
\alias{rasterlists,Seurat-method}
\alias{rasterlists,Seurat,Seurat-method}
\title{Extracts available image types from Seurat or Staffli objects}
\usage{
rasterlists(object)

rasterlists(object)

\S4method{rasterlists}{Staffli}(object)

\S4method{rasterlists}{Seurat}(object)
}
\arguments{
\item{object}{A Staffli or Seurat object}
}
\description{
Extracts available image types from Seurat or Staffli objects

Method to extract the available image types from a Staffli object
}
