% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rigid_transformation.R
\name{rigid.refl}
\alias{rigid.refl}
\title{Creates a transformation matrix for reflection}
\usage{
rigid.refl(mirror.x, mirror.y)
}
\arguments{
\item{mirror.x, mirror.y}{Logical: mirrors x or y axis if set to TRUE}
}
\description{
Creates a transformation matrix for reflection where mirror.x will reflect the
points along the x axis and mirror.y will reflect thepoints along the y axis.
Points are assumed to be centered at (0, 0)
}
