% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rigid_transformation.R
\name{rigid.stretch}
\alias{rigid.stretch}
\title{Creates a transformation matrix for stretching}
\usage{
rigid.stretch(r)
}
\arguments{
\item{r}{stretching factor}
}
\description{
Creates a transformation matrix for stretching by a factor of r 
along the x axis.
}
