% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rigid_transformation.R
\name{rigid.transl}
\alias{rigid.transl}
\title{Creates a transformation matrix for translation with an offset of (h, k)}
\usage{
rigid.transl(h = 0, k = 0)
}
\arguments{
\item{h}{Numeric: offset along x axis}

\item{k}{Numeric: offset along y axis}
}
\description{
Creates a transformation matrix for translation with an offset of (h, k)
}
