% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunNMF.R
\name{rnmf}
\alias{rnmf}
\title{Run NMF with ica init}
\usage{
rnmf(
  A,
  k,
  alpha = 0,
  init = "ica",
  n.cores = 1,
  loss = "mse",
  max.iter = 500,
  ica.fast = F
)
}
\arguments{
\item{A}{Expression matrix}

\item{k}{= Number of factors}

\item{alpha}{Alpha parameter value}

\item{init}{Method to initiate NMF by}

\item{n.cores}{Number of threads to run}

\item{loss}{Loss method}

\item{max.iter}{Maximum number of iterations for NMF}

\item{ica.fast}{Should fast ica be run?}
}
\description{
Run NMF with ica init
}
