% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rigid_transformation.R
\name{rotate}
\alias{rotate}
\title{Creates a transformation matrix that rotates an object
in 2D around it's center of mass}
\usage{
rotate(angle, center.cur)
}
\arguments{
\item{angle}{Angle given in degrees used for rotation}

\item{center.cur}{Coordinates of the current center of mass}
}
\description{
Creates a transformation matrix that rotates an object
in 2D around it's center of mass
}
