% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunNMF.R
\name{RunNMF}
\alias{RunNMF}
\title{Run Non-negative Matrix Factorization}
\usage{
RunNMF(object, assay = NULL, features = NULL, nfactors = 20,
  ndims.print = 1:5, nfeatures.print = 30, reduction.name = "NMF",
  reduction.key = "factor_", n.cores = 7, ...)
}
\arguments{
\item{object}{Seurat object}

\item{assay}{assay Name of Assay PCA is being run on}

\item{features}{features to compute the NMF for}

\item{nfactors}{Total Number of factors to compute and store (20 by default)}

\item{nfeatures.print}{Number of genes to print for each factor}

\item{reduction.name}{dimensional reduction name,  pca by default}

\item{reduction.key}{dimensional reduction key, specifies the string before
the number for the dimension names. factor by default}

\item{nfactors.print}{factorss to print genes for}
}
\description{
Run Non-negative Matrix Factorization
}
