% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STUtility.R, R/hello.R
\name{runSEH}
\alias{runSEH}
\title{Spatial Expression Histology}
\usage{
runSEH(se, n.clust = 10, sample.wise.z = FALSE, global.z = FALSE,
  m.var = 0, log.freq = TRUE, use.cov = TRUE, n.start = 10,
  m.iter = 10, db.scan = FALSE, db.sca.eps = 0.15,
  db.scan.MinPts = 5)

runSEH(se, n.clust = 10, sample.wise.z = FALSE, global.z = FALSE,
  m.var = 0, log.freq = TRUE, use.cov = TRUE, n.start = 10,
  m.iter = 10, db.scan = FALSE, db.sca.eps = 0.15,
  db.scan.MinPts = 5)
}
\description{
Spatial Expression Histology

Spatial Expression Histology
}
