% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Staffli.R
\docType{methods}
\name{samplenames,Seurat-method}
\alias{samplenames,Seurat-method}
\title{Method used to get samplenames from a Staffli object}
\usage{
\S4method{samplenames}{Seurat}(object)
}
\arguments{
\item{A}{Seurat object}
}
\description{
Method used to get samplenames from a Staffli object
}
