% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/Staffli.R
\docType{methods}
\name{samplenames}
\alias{samplenames}
\alias{samplenames,Staffli-method}
\alias{samplenames,Staffli,Staffli-method}
\alias{samplenames,Seurat-method}
\alias{samplenames,Seurat,Seurat-method}
\title{Extracts available samplenames from Seurat or Staffli objects}
\usage{
samplenames(object)

samplenames(object)

\S4method{samplenames}{Staffli}(object)

\S4method{samplenames}{Seurat}(object)
}
\arguments{
\item{object}{A Staffli object}
}
\description{
Extracts available samplenames from Seurat or Staffli objects

Method used to get samplenames from a Staffli or Seurat object processed using STutility
}
