% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_grid_SC.R
\name{scale_override}
\alias{scale_override}
\title{Function used to override scales in facets of a ggplot object}
\usage{
scale_override(which, scale)
}
\arguments{
\item{which}{Index of facet}

\item{scale}{Object of class "Scale" used to override (e.g. \code{scale_x_continuous()})}
}
\description{
Function used to override scales in facets of a ggplot object
}
