% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Staffli.R
\docType{methods}
\name{scaled.imdims,Staffli-method}
\alias{scaled.imdims,Staffli-method}
\title{Method for extracting the scaled HE image dimensions from a 'Staffli' object}
\usage{
\S4method{scaled.imdims}{Staffli}(object, type = "raw")
}
\arguments{
\item{object}{A Staffli object}

\item{type}{Image type [choices: 'raw', 'masked', 'processed']}
}
\description{
Method for extracting the scaled HE image dimensions from a 'Staffli' object
}
