% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{scaled.imdims}
\alias{scaled.imdims}
\title{Extracts scaled dimensions of images in Staffli object}
\usage{
scaled.imdims(object, type = "raw")
}
\arguments{
\item{object}{A Staffli object}
}
\description{
Extracts scaled dimensions of images in Staffli object
}
