% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D_point_pattern.R
\name{scatter_HE}
\alias{scatter_HE}
\title{Function used to generate a 2D set of points from an HE image}
\usage{
scatter_HE(
  object,
  type = "masked.masks",
  sample.index = NULL,
  limit = 0.5,
  maxnum = 50000,
  edges = FALSE,
  custom.edge.detector = NULL
)
}
\arguments{
\item{object}{Staffli object}

\item{type}{Sets the image type to run segmentation on}

\item{sample.index}{Integer value specifying the index of the sample to be analyzed}

\item{limit}{Sets the intensity threshold in the interval [0, 1]}

\item{maxnum}{Integer value specifying the maximum number of points}

\item{edges}{Extracts the coordinates of the edges instead}

\item{custom.edge.detector}{Custom function used to detect edges}
}
\description{
The defualt segmentation uses the HE image as input and defines any pixel with an intensity value
below a threshold to be a point. The number of points can be downsampeld to limit the maximum number.
}
