% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STUtility.R, R/hello.R
\name{seh.plot}
\alias{seh.plot}
\title{Plot spatial expression histology heatmap}
\usage{
seh.plot(se, filename, bg = "black", col.scale.cluster = TRUE,
  disable.voronoi = FALSE, marBot = 0, marLeft = 0, marRight = 0,
  marTop = 0)

seh.plot(se, filename, bg = "black", col.scale.cluster = TRUE,
  disable.voronoi = FALSE, marBot = 0, marLeft = 0, marRight = 0,
  marTop = 0)
}
\arguments{
\item{se}{Summarized experiment.}

\item{filename}{Name of output file.}

\item{bg}{Backsampound color.}

\item{col.scale.cluster}{Colors will be scaled per gene cluster across all samples.}

\item{disable.voronoi}{Plot points instead of voronoi}

\item{marBot}{Bottom plot margin}

\item{marLeft}{Left plot margin}

\item{marRight}{Right plot margin}

\item{marTop}{Top plot margin}

\item{se}{Summarized experiment.}

\item{filename}{Name of output file.}

\item{bg}{Backsampound color.}

\item{col.scale.cluster}{Colors will be scaled per gene cluster across all samples.}

\item{disable.voronoi}{Plot points instead of voronoi}

\item{marBot}{Bottom plot margin}

\item{marLeft}{Left plot margin}

\item{marRight}{Right plot margin}

\item{marTop}{Top plot margin}
}
\description{
Plot spatial expression histology heatmap

Plot spatial expression histology heatmap
}
