% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEH.R
\name{seh.tsne}
\alias{seh.tsne}
\title{Create tSNE based on spatial expression histology}
\usage{
seh.tsne(se, filename, perplexity = 30, priorPCA = FALSE,
  bg = "black")
}
\arguments{
\item{se}{s4 object}

\item{priorPCA}{use PCA as input to tSNE}

\item{bg}{plot background color}

\item{perpelxity}{perplexity paramter tSNE}
}
\description{
Create tSNE based on spatial expression histology
}
