% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_utilities.R
\name{SetQuantile}
\alias{SetQuantile}
\title{Find the quantile of a data}
\usage{
SetQuantile(cutoff, data)
}
\arguments{
\item{cutoff}{The cutoff to turn into a quantile}

\item{data}{The data to find the quantile of}
}
\value{
The numerical representation of the quantile
}
\description{
Converts a quantile in character form to a number regarding some data
String form for a quantile is represented as a number prefixed with 'q'
For example, 10th quantile is 'q10' while 2nd quantile is 'q2'
}
\details{
Will only take a quantile of non-zero data values
}
\references{
\url{https://github.com/satijalab/seurat/blob/master/R/visualization.R}
}
