% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Staffli.R
\docType{methods}
\name{show,Staffli-method}
\alias{show,Staffli-method}
\title{Show metho for Staffli objects}
\usage{
\S4method{show}{Staffli}(object)
}
\arguments{
\item{object}{A Staffli object}
}
\description{
Show metho for Staffli objects
}
