% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_processing_utilities.R
\name{sleepy}
\alias{sleepy}
\title{Function to perform system sleep if user having memory issues during image loading}
\usage{
sleepy(x)
}
\arguments{
\item{x}{time to sleep}
}
\description{
Function to perform system sleep if user having memory issues during image loading
}
\keyword{internal}
