% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_processing_utilities.R
\name{slic}
\alias{slic}
\title{SLIC algorithm}
\usage{
slic(im, nS, compactness = 1, ...)
}
\arguments{
\item{im}{Image of class "cimg"}

\item{nS}{Number of superpixels to return}

\item{compactness}{Controls scaling ratio for pixel values}

\item{...}{Parameters passed to kmeans}
}
\description{
Converts image into SLIC superpixels
}
