% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{spatial_dim_plot}
\alias{spatial_dim_plot}
\title{Dimensional reduction plot on ST coordinates on top of HE image}
\usage{
spatial_dim_plot(object, dims = 1:2, sample.index = 1, spots = NULL,
  type = NULL, min.cutoff = NA, max.cutoff = NA, blend = FALSE,
  pt.size = 1, pt.alpha = 1, pt.border = FALSE, add.alpha = FALSE,
  reduction = NULL, shape.by = NULL, palette = "MaYl", cols = NULL,
  grid.ncol = NULL, center.zero = TRUE, channels.use = NULL,
  verbose = FALSE, dark.theme = FALSE, sample.label = TRUE,
  show.sb = TRUE, value.scale = c("samplewise", "all"),
  custom.theme = NULL, ...)
}
\arguments{
\item{object}{Seurat object}

\item{dims}{List of dimensions for original images. This list has to contain one element for each sample and each element
should be a vector of length 2 specifying the dimensions of the original HE image.}

\item{sample.index}{Index specifying the sample that you want to use for plotting}

\item{spots}{Character vector with spot IDs to plot [default: all spots]}

\item{type}{Image type to plot on. Here you can specify any of the images available in your Seurat object. To get this list you can
run the \code{\link{rasterlists}} function on your Seurat object. If the type is not specified, the images will be prioritized in the following
order if they are available; "processed", "masked" and "raw".}

\item{min.cutoff}{Vector of minimum and maximum cutoff values for each feature, may specify quantile in the form of 'q##' where '##'
is the quantile (eg, 'q1', 'q10'). This can be useful if you have outlier values that skew the colorscale in the plot. For example, if you specify
'q1', you will trim of values below the 1st percentile. [default: no cuttoffs]}

\item{max.cutoff}{Vector of minimum and maximum cutoff values for each feature, may specify quantile in the form of 'q##' where '##'
is the quantile (eg, 'q1', 'q10'). This can be useful if you have outlier values that skew the colorscale in the plot. For example, if you specify
'q1', you will trim of values below the 1st percentile. [default: no cuttoffs]}

\item{blend}{Scale and blend expression values to visualize coexpression of two features (this options will override other coloring parameters).
See 'Blending values' below for a more thourough description.}

\item{pt.size}{Point size of each ST spot [default: 1]}

\item{pt.alpha}{Opacity of each ST spot [default: 1]}

\item{pt.border}{Should a border be drawn around the spots? [default: TRUE]}

\item{reduction}{Which dimensionality reduction to use. If not specified, first searches for "umap", then "tsne", then "pca"}

\item{shape.by}{String specifying the column where the shaping label is stored}

\item{palette}{Color palette used for spatial heatmap (see \code{palette.select(info = T)} for available options).
Disabled if a color vector is provided (see \code{cols} below).}

\item{cols}{A vector of colors to use for colorscale, e.g. \code{cols = c("blue", "white", "red")} will
create a gradient color scale going from blue to white to red. This options will deactivate the \code{palette}
option.}

\item{grid.ncol}{Number of columns for display when combining plots. This option will only have an effect on the sample level structure.}

\item{center.zero}{Specifies whther or not the colorscale should be centered around 0. For some values, such as Principal Component vectors,
the distribution of values is centered at 0 and in that case it can be appropriate to use a divergent colorscale with a predefined value for 0.
If this parameter is set to TRUE, the ggplot2 function \code{scale_color_gradient2} will be used to control the coloring instead of
\code{scale_color_gradientn}. If center.zero is set to FALSE, the colorscale will simply map the values in equally spaced intervals which could skew
the interpretaion of the output plot.}

\item{channels.use}{Color channels to use for blending. Has to be a character vector of length 2 or 3 with "red", "green" and "blue"
color names specified [default: c("red", "green", "blue)]}

\item{verbose}{Print messages}

\item{dark.theme}{Switches color of scalebar to 'white'}

\item{sample.label}{Should the sample label be included in the image? [default: TRUE]}

\item{show.sb}{Should the size bar be displayed? [default: TRUE]}

\item{value.scale}{Defines how the feature values should be mapped to the colorbar. If `value.scale = "samplewise"`, each feature will be
scaled independently and if `value.scale = "all"` the features will all have the same value reange.}

\item{custom.theme}{Object of class 'theme' used to change the background theme (see for example \url{https://ggplot2.tidyverse.org/reference/theme.html})}

\item{...}{Extra parameters passed on to \code{\link{ST.ImagePlot}}}
}
\value{
A ggplot object
}
\description{
Dimensional reduction plot on ST coordinates on top of HE image
}
