% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_genes.R
\name{SpatialGenes}
\alias{SpatialGenes}
\title{Find genes with spatial structure}
\usage{
SpatialGenes(object, spots = NULL, slot = "data",
  metric = "euclidean", features.include = NULL, rbp_p = 0.95,
  examine_top = 0.3, python_path = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{spots}{Character vector with spot IDs to plot [default: all spots]}

\item{slot}{Which slot to pull the data from? [default: 'data']}

\item{metric}{The distance method used for `scipy.spatial.distance.pdist`
The distance function can be ‘braycurtis’, ‘canberra’, ‘chebyshev’, ‘cityblock’,
‘correlation’, ‘cosine’, ‘dice’, ‘euclidean’, ‘hamming’, ‘jaccard’,
‘jensenshannon’, ‘kulsinski’, ‘mahalanobis’, ‘matching’, ‘minkowski’,
‘rogerstanimoto’, ‘russellrao’, ‘seuclidean’, ‘sokalmichener’,
‘sokalsneath’, ‘sqeuclidean’ or ‘yule’.}

\item{features.include}{Features to include in the detection computation}

\item{rbp_p}{Fractional binarization method}

\item{examine_top}{Top fraction to evaluate with silhouette}

\item{python_path}{Specify specific path to python if required}
}
\value{
list of data.frames with spatial scores for each gene
}
\description{
Runs a distance based method to detect genes with spatial structure.
}
\section{Acknowledgement}{

The function is based on the `calculate_spatial_genes_python()` available from
the `Giotto` R package, but have been modified to work with Seurat objects
and on multiple samples.
}

\examples{
spatgenes <- SpatialGenes(se)

}
