% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_LR.R
\name{SpatialLR}
\alias{SpatialLR}
\title{Detect ligand-receptor interactions}
\usage{
SpatialLR(
  object,
  LR,
  method = "binarize",
  slot = "data",
  smooth.data = TRUE,
  group.by = NULL,
  nNeighbors = 20,
  return.interaction.assay = FALSE,
  return.smooth.assay = FALSE,
  do.parallel = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{object}{Seurat object}

\item{LR}{A data.frame with ligand-receptor pairs. Columns should be named
"ligand" and "receptor".}

\item{method}{"binarize" or "lee". See details for more information.}

\item{smooth.data}{Logical indicating whether or not the data should be smoothened
or not. Smoothing of expression data is used by averaging the expression across
neighboring spots.}

\item{nNeighbors}{Number of neighbors to use for smoothing. More neighbors will
lead to more smoothing of the data.}

\item{return.interaction.assay}{Returns an assay with the interactions.}

\item{do.parallel}{Run the function on multiple cores. The number of cores can be
set by the MC_CORES envornment variable, e.g:
`Sys.setenv("MC_CORES" = 7L)`}

\item{verbose}{Print messages}
}
\value{
data.frame with ligand-receptor interactions
}
\description{
This function can be used to find ligand-receptor pairs that are expressed
in the same area of the tissue.
}
