% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{st.load.matrix}
\alias{st.load.matrix}
\title{load count files}
\usage{
st.load.matrix(path, delim = "\\t", row.names = 1, visium = F)
}
\arguments{
\item{path}{Path to count files}

\item{delim}{Delimiter used in gene count matrix file}

\item{row.names}{set column as row.names}

\item{visium}{Load 10x visium output files}
}
\description{
load count files
}
\keyword{internal}
