% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_utilities.R
\name{ST.rgbDimPlot}
\alias{ST.rgbDimPlot}
\title{Store rgb values after ST.Dimplot
OBS! could be merged with ST.Dimplot?}
\usage{
ST.rgbDimPlot(plotObj)
}
\arguments{
\item{plotObj}{A ST.DimPlot output object}
}
\description{
Store rgb values after ST.Dimplot
OBS! could be merged with ST.Dimplot?
}
