% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEH.R
\name{ST.seh}
\alias{ST.seh}
\title{==============  Spatial Expression Histology ====================================}
\usage{
ST.seh(se, n.clust = 10, n.top = 0, log.freq = TRUE, n.start = 10,
  n.iter = 10, use.moduleScore = FALSE, use.scale = TRUE, ...)
}
\arguments{
\item{se}{s4 object}

\item{n.clust}{number of clusters}

\item{log.freq}{use log of freq}

\item{n.start}{number of starts kMeans}

\item{sample.wise.z}{sample wize z-score calculation}

\item{global.z}{global z-score calculation}

\item{m.var}{variance?}

\item{use.cov}{use covarance}

\item{m.iter}{number of iterations kMeans}
}
\description{
Covariance based spatial expression Histology
}
