% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Staffli.R
\name{[[,Staffli,ANY,ANY-method}
\alias{[[,Staffli,ANY,ANY-method}
\alias{`[[`,Staffli,Staffli-method}
\alias{[,Staffli,ANY,ANY,ANY-method}
\alias{`[`,Staffli,Staffli-method}
\title{Method used to get meta data from a Staffli object}
\usage{
\S4method{[[}{Staffli,ANY,ANY}(x, i, j, drop = F)

\S4method{[}{Staffli,ANY,ANY,ANY}(x, i)
}
\arguments{
\item{x}{Staffli bject from which to get image from.}

\item{i}{index for image to get.}

\item{j}{column indices specifying elements to extract.}

\item{drop}{If TRUE the result is coerced to the lowest possible dimension. 
This only works for extracting elements, not for the replacement.}
}
\description{
Method used to get meta data from a Staffli object

Method used to get list of scaled images
from selected type
}
