% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Staffli.R
\name{[[<-,Staffli,ANY,ANY-method}
\alias{[[<-,Staffli,ANY,ANY-method}
\alias{`[[<-`,Staffli,Staffli-method}
\alias{[<-,Staffli,ANY,ANY,ANY-method}
\alias{`[<-`,Staffli,Staffli-method}
\title{Method used to set meta data in a Staffli object}
\usage{
\S4method{[[}{Staffli,ANY,ANY}(x, i, j, ...) <- value

\S4method{[}{Staffli,ANY,ANY,ANY}(x, i, j, ...) <- value
}
\arguments{
\item{x}{Staffli object in which to replace image.}

\item{i}{index for image to replace.}

\item{j}{column indices specifying elements to replace.}

\item{...}{additional parameters}

\item{value}{Data to add to meta data data.frame.}
}
\description{
Method used to set meta data in a Staffli object

Method used to set list of scaled images
of selected type
}
