% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rigid_transformation.R
\name{stretch}
\alias{stretch}
\title{Stretch along angle}
\usage{
stretch(r, alpha, center.cur)
}
\arguments{
\item{r}{stretching factor}

\item{alpha}{angle}

\item{center.cur}{Coordinates of the current center of mass}
}
\description{
Creates a transformation matrix that stretches an object
along a specific axis
}
