% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R, R/manual_annotation.R,
%   R/pre_processing.R, R/process_from_table.R
\docType{package}
\name{STutility-package}
\alias{STutility}
\alias{STutility-package}
\title{STutility: Analysis and visualization of Spatial Transcriptomics data}
\description{
STutility collects useful tools for Spatial Transcriptomics data analysis within R. The package is built on top of popular Seurat R package, which is a toolkit for analysis and exploration of single cell RNA sequencing data (scRNAseq). STutility adds image functionallity to the Seurat workflow, allowing RNAseq data to be put in a spatial tissue context. Tutorials for data transformation, QC, analysis and visualization can be found on our website: https://ludvigla.github.io/STUtility_web_site/.

STutility collects useful tools for Spatial Transcriptomics data analysis within R. The package is built on top of popular Seurat R package, which is a toolkit for analysis and exploration of single cell RNA sequencing data (scRNAseq). STutility adds image functionallity to the Seurat workflow, allowing RNAseq data to be put in a spatial tissue context. Tutorials for data transformation, QC, analysis and visualization can be found on our website: https://ludvigla.github.io/STUtility_web_site/.

STutility collects useful tools for Spatial Transcriptomics data analysis within R. The package is built on top of popular Seurat R package, which is a toolkit for analysis and exploration of single cell RNA sequencing data (scRNAseq). STutility adds image functionallity to the Seurat workflow, allowing RNAseq data to be put in a spatial tissue context. Tutorials for data transformation, QC, analysis and visualization can be found on our website: https://ludvigla.github.io/STUtility_web_site/.

STutility collects useful tools for Spatial Transcriptomics data analysis within R. The package is built on top of popular Seurat R package, which is a toolkit for analysis and exploration of single cell RNA sequencing data (scRNAseq). STutility adds image functionallity to the Seurat workflow, allowing RNAseq data to be put in a spatial tissue context. Tutorials for data transformation, QC, analysis and visualization can be found on our website: https://ludvigla.github.io/STUtility_web_site/.
}
\author{
\strong{Maintainer}: Ludvig Larsson \email{ludvig.larsson@scilifelab.se}

Authors:
\itemize{
  \item Joseph Bergenstraahle
}


\strong{Maintainer}: Ludvig Larsson \email{ludvig.larsson@scilifelab.se}

Authors:
\itemize{
  \item Joseph Bergenstraahle
}


\strong{Maintainer}: Ludvig Larsson \email{ludvig.larsson@scilifelab.se}

Authors:
\itemize{
  \item Joseph Bergenstraahle
}


\strong{Maintainer}: Ludvig Larsson \email{ludvig.larsson@scilifelab.se}

Authors:
\itemize{
  \item Joseph Bergenstraahle
}

}
