% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Staffli.R
\name{[[,Staffli,ANY,ANY-method}
\alias{[[,Staffli,ANY,ANY-method}
\title{Method used to get meta data from a Staffli object}
\usage{
\S4method{[[}{Staffli,ANY,ANY}(x, i, j, drop = F)
}
\arguments{
\item{x}{A Staffli object}

\item{i}{Row names or indices to subset x on}

\item{j}{Column names or indices to subset x on}
}
\description{
Method used to get meta data from a Staffli object
}
