% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Staffli.R
\name{[[<-,Staffli,ANY,ANY,ANY-method}
\alias{[[<-,Staffli,ANY,ANY,ANY-method}
\title{Method used to set meta data in a Staffli object}
\usage{
\S4method{[[}{Staffli,ANY,ANY,ANY}(x, i, j, ...) <- value
}
\arguments{
\item{x}{A Staffli object}

\item{i}{Row names or indices to subset x on}

\item{j}{Column names or indices to subset x on}

\item{value}{Value to save into Staffli meta data}
}
\description{
Method used to set meta data in a Staffli object
}
