% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Staffli.R
\docType{methods}
\name{[<-,Staffli,ANY,ANY,ANY-method}
\alias{[<-,Staffli,ANY,ANY,ANY-method}
\title{Method used to set list of scaled images
of selected type}
\usage{
\S4method{[}{Staffli,ANY,ANY,ANY}(x, i, j, ...) <- value
}
\arguments{
\item{x}{A Staffli object}

\item{i}{A characetr string specifying the name of the
images to set; "raw", "masked" or "processed"}
}
\description{
Method used to set list of scaled images
of selected type
}
