% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Staffli.R
\name{SubsetSTData}
\alias{SubsetSTData}
\title{Subset a Seurat objectcontaining Staffli image data}
\usage{
SubsetSTData(object, expression, spots = NULL, features = NULL,
  idents = NULL, ...)
}
\arguments{
\item{expression}{Logical expression indicating features/variables to keep}

\item{spots}{A vector of spots to keep}

\item{features}{A vector of features to keep}

\item{idents}{A vector of identity classes to keep}

\item{...}{Extra parameters passed to WhichCells, such as slot, invert, or downsample}
}
\description{
Subsets a Seurat object containing Spatial Transcriptomics data while
making sure that the images are subsetted correctly. If you use the default
\code{\link{subset}} function you will not be able to use any of the
STUtility visualization methods.
}
