% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunNMF.R
\name{SummarizeAssocFeatures}
\alias{SummarizeAssocFeatures}
\title{Summarize features associated with certain factors}
\usage{
SummarizeAssocFeatures(object, features.return = 10,
  features.use = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{features.return}{Number of features to return per factor}

\item{features.use}{Select features to use for summary}
}
\description{
Summarize features associated with certain factors
}
