% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rigid_transformation.R
\name{translate}
\alias{translate}
\title{Creates a transformation matrix that translates an object
in 2D}
\usage{
translate(translate.x, translate.y)
}
\arguments{
\item{translate.x, translate.y}{translation of x, y coordinates}
}
\description{
Creates a transformation matrix that translates an object
in 2D
}
