% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/image_processing.R
\name{WarpImages}
\alias{WarpImages}
\alias{WarpImages.Staffli}
\alias{WarpImages.Seurat}
\title{Warps images using various transformations}
\usage{
WarpImages(object, transforms, verbose)

\method{WarpImages}{Staffli}(object, transforms, verbose = FALSE)

\method{WarpImages}{Seurat}(object, transforms, verbose = FALSE)
}
\arguments{
\item{object}{Seurat object}

\item{transforms}{List of arguments passed to warp function}

\item{verbose}{Print messages}
}
\value{
Seurat object with processed imaged

A Staffli object

A Seurat object
}
\description{
Warps images using various transformations
}
\examples{
# Create a new Staffli object, mask, warp and plot images
st.obj <- CreateStaffliObject(imgs, meta.data)
transforms <- list("2" = list("mirror.y" = TRUE))
st.obj <- LoadImages(st.obj, verbose = TRUE) \%>\% MaskImages() \%>\% WarpImages(transforms)
plot(st.obj)
# Load, mask, warp and plot images in a Seurat object
# Mirror y axis in sample '2' and rotate sample '3' 10 degrees
transforms <- list("2" = list("mirror.y" = TRUE), "3" = list("angle" = 10))
se <- LoadImages(se, verbose = TRUE) \%>\% MaskImages() \%>\% WarpImages(transforms)
ImagePlot(se)
}
