% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_genes.R
\name{CorSpatialGenes}
\alias{CorSpatialGenes}
\title{Orders spatially correlated genes}
\usage{
CorSpatialGenes(object, assay = NULL, slot = "scale.data",
  features = NULL, nNeighbours = NULL, maxdist = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{assay}{Name of assay the function is being run on}

\item{slot}{Slot to use as input [default: 'scale.data']}

\item{features}{Features to rank}

\item{nNeighbours}{Number of neighbours to find for each spot}

\item{maxdist}{Maximum allowed distance to define neighbouring spots [default: 1.5]}
}
\value{
data.frame with gene names and correlation scores
}
\description{
This function can be used to find genes with spatial structure in ST data
by using lag vectors.
}
\details{
\itemize{
   \item{Builds a connection network from the array x,y coordinates for each sample. For a 'Visium' array, this would typically be 6 neighbours
   because of the hexagonal structure of spots.}
   \item{Combine connection networks from multiple samples}
   \item{Compute the lag vector for each feature}
   \item{Compute the correlation between the lag vector and the original vector}
}
}
