% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3D_viusalization.R
\name{Create3DStack}
\alias{Create3DStack}
\title{Creates a stack of point patterns with aligned data and stores it in a
`Staffli` object.}
\usage{
Create3DStack(object, limit = 0.4, maxnum = 50000, nx = 200, verbose = TRUE)
}
\arguments{
\item{object}{Seurat object}

\item{limit}{Cut-off threshold for segmentation of points. Has to be a value between 0-1.}

\item{maxnum}{Maximum number of points to store for each section [default: 5e5]. If you are analyzing many samples
you will have to decrease this number.}

\item{nx}{Number of cells to interpolate over across the x axis}

\item{verbose}{Print messages}
}
\value{
Seurat object
}
\description{
Creates a stack of point patterns with aligned data and stores it in a
`Staffli` object.
}
