% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{DimOverlay}
\alias{DimOverlay}
\title{Dimensional reduction plot on ST coordinates on top of HE image}
\usage{
DimOverlay(object, dims = 1:2, sample.index = 1, type = NULL,
  min.cutoff = NA, max.cutoff = NA, blend = FALSE, pt.size = 1,
  pt.alpha = 1, reduction = NULL, shape.by = NULL, palette = NULL,
  cols = NULL, rev.cols = FALSE, grid.ncol = NULL,
  center.zero = TRUE, channels.use = NULL, verbose = FALSE,
  dark.theme = FALSE, ...)
}
\arguments{
\item{object}{Seurat object}

\item{dims}{List of dimensions for original images. This list has to contain one element for each sample and each element
should be a vector of length 2 specifying the dimensions of the original HE image.}

\item{sample.index}{Index specifying the sample that you want to use for plotting}

\item{type}{Image type to plot on. Here you can specify any of the images available in your Seurat object. To get this list you can
run the \code{\link{rasterlists}} function on your Seurat object. If the type is not specified, the images will be prioritized in the following
order if they are available; "processed", "masked" and "raw".}

\item{min.cutoff}{Vector of minimum and maximum cutoff values for each feature, may specify quantile in the form of 'q##' where '##'
is the quantile (eg, 'q1', 'q10'). This can be useful if you have outlier values that skew the colorscale in the plot. For example, if you specify
'q1', you will trim of values below the 1st percentile. [default: no cuttoffs]}

\item{max.cutoff}{Vector of minimum and maximum cutoff values for each feature, may specify quantile in the form of 'q##' where '##'
is the quantile (eg, 'q1', 'q10'). This can be useful if you have outlier values that skew the colorscale in the plot. For example, if you specify
'q1', you will trim of values below the 1st percentile. [default: no cuttoffs]}

\item{blend}{Scale and blend expression values to visualize coexpression of two features (this options will override other coloring parameters).
See 'Blending values' below for a more thourough description.}

\item{pt.size}{Point size of each ST spot [default: 1]}

\item{pt.alpha}{Opacity of each ST spot [default: 1]}

\item{reduction}{Which dimensionality reduction to use. If not specified, first searches for "umap", then "tsne", then "pca"}

\item{shape.by}{String specifying the column where the shaping label is stored}

\item{palette}{Color palette used for spatial heatmap (see \code{palette.select(info = T)} for available options).
Disabled if a color vector is provided (see \code{cols} below).}

\item{cols}{A vector of colors to use for colorscale, e.g. \code{cols = c("blue", "white", "red")} will
create a gradient color scale going from blue to white to red. This options will deactivate the \code{palette}
option.}

\item{rev.cols}{Logical specifying whether colorscale should be reversed [default: FALSE]}

\item{grid.ncol}{Number of columns for display when combining plots. This option will only have an effect on the sample level structure.}

\item{center.zero}{Specifies whther or not the colorscale should be centered around 0. For some values, such as Principal Component vectors,
the distribution of values is centered at 0 and in that case it can be appropriate to use a divergent colorscale with a predefined value for 0.
If this parameter is set to TRUE, the ggplot2 function \code{scale_color_gradient2} will be used to control the coloring instead of
\code{scale_color_gradientn}. If center.zero is set to FALSE, the colorscale will simply map the values in equally spaced intervals which could skew
the interpretaion of the output plot.}

\item{verbose}{Print messages}

\item{...}{Extra parameters passed on to \code{\link{ST.ImagePlot}}}

\item{slot}{Which slot to pull expression data from? [default: 'data']}
}
\value{
A ggplot object
}
\description{
Graphs the selected vectors of a dimensional reduction technique on a 2D grid of spots overlaid on top of an HE images.
}
\details{
Note that this function only allows you to plot one sample at the time. It is typically difficult to explore details in the HE image
when diplaying multiple samples side by side, so we decided to limit the choice to plot one sample at the time. If you have higher resolution images,
it could also take significant time to draw the plots. If you still wish to show multiple samples together you can use the \code{\link{MultiFeatureOverlay}}
function.
}
\section{Blending values}{

The blend option can be useful if you wish to visualize multiple dimensions simultaneuosly and works for two or three dimensionality
reduction vectors. Each of the selected vectors are rescaled from 0 to 1 and are used as RGB color channels to produce mixed color for each
spot. This can be particularly useful when looking at overlapping value vectors. For example, if you are looking at two overlapping value vectors
"A" and "B" and use the blend option, the "A" values will be encoded in the "red" channel and the "B" values in the "green" channel. If a spot is
purely "A" the color will be red and if it is purely "B" it will green. Any mixture of "A" and "B" will produce a color between red and green
where a 50/50 mixture gives yellow color. The amplitude if the values will also determine the brightness of the color.
}

\examples{

# Load images and run PCA
se <- LoadImages(se) \%>\%
   RunPCA()

# Plot the first 2 dimensions
DimOverlay(se, dims = 1:2, reduction = "pca", sample.index = 1)

# Blend values for dimensions 1 and 2
DimOverlay(se, dims = 1:2, reduction = "pca", sample.index = 1, blend = T)

# Plot the first 2 dimensions and trim off 1st percentile values
DimOverlay(se, dims = 1:2, reduction = "pca", sample.index = 1, min.cutoff = 'q1')

# Mask images and plot plot the first 2 dimensions on the masked image
se <- MaskImages(se)
DimOverlay(se, dims = 1:2, reduction = "pca", sample.index = 1, type = "masked")

}
\seealso{
\code{\link{ST.FeaturePlot}} and \code{\link{ST.DimPlot}} for how to plot features
without the HE image, \code{\link{FeatureOverlay}} for how to overlay feature plots on the
HE images and \code{\link{MultiFeatureOverlay}} and \code{\link{MultiDimOverlay}} for how
to overlay plots on the HE images in multiple samples.
}
