% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_utilities.R
\name{draw_scalebar}
\alias{draw_scalebar}
\title{Draws a scalebar for ST plots}
\usage{
draw_scalebar(p, x, xend, y)
}
\arguments{
\item{p}{An object of class 'ggplot'}

\item{x, xend}{The start and positions of the scalebar along the x axis.}

\item{y}{The position of the scalebar along the y axis.}
}
\value{
An object of class 'ggplot' with a scalebar drawn on top of it
}
\description{
Takes a ggplot object as input together with some additional parameters
specifying the position and widht of a scalebar representing the actual
width of corresponding 500um in the plot. The actual width of the scale
bar is determined by the x, xend parameters and has to be calulated
separately. To do this, you can use the pixels.per.um slot of the
`Staffli` object.
}
