% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{feature.scaler}
\alias{feature.scaler}
\title{Function used to scale numerical features}
\usage{
feature.scaler(data, features, min.cutoff, max.cutoff, spots)
}
\arguments{
\item{data}{data.frame containing x, y coordinates and columns with numerical features}

\item{min.cutoff, max.cutoff}{Vector of minimum and maximum cutoff values for each feature,
may specify quantile in the form of 'q##' where '##' is the quantile (eg, 'q1', 'q10')}
}
\description{
Function used to scale numerical features
}
