% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{g_legend}
\alias{g_legend}
\title{Create legend}
\usage{
g_legend(data, variable, center.zero, cols, val.limits)
}
\arguments{
\item{data}{data.frame with values that should be mapped onto colorscale}

\item{variable}{character string specifying tha name of the column (variable) containing values}

\item{center.zero}{Specifies whether or not the colorscale should be centered at zero}

\item{cols}{Character vector with color ids to be used in colorscale}

\item{val.limits}{Specifies the limits for values in colorscale}
}
\description{
Creates a colorscale legend for a dataset in grob format
}
