% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_processing_utilities.R
\name{get.edges.Staffli}
\alias{get.edges.Staffli}
\title{Finds edges of a mask and returns a set of x,y coordinates
aligned along the edges}
\usage{
\method{get.edges}{Staffli}(object, index = 1, verbose = FALSE,
  type = "masked.masks")
}
\arguments{
\item{object}{Seurat object}

\item{index}{Sample index}

\item{verbose}{Print messages}
}
\value{
A list of data.frames with edge coordinates
}
\description{
Finds edges of a mask and returns a set of x,y coordinates
aligned along the edges
}
\examples{
# Create a new Staffli object, mask, align and plot edges for image 2
st.obj <- CreateStaffliObject(imgs, meta.data)
edges <- LoadImages(st.obj, verbose = TRUE) \%>\% MaskImages() \%>\% get.edges(index = 2)
plot(as.raster(edges) \%>\% as.cimg())
}
