% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/Staffli.R
\docType{methods}
\name{iminfo}
\alias{iminfo}
\alias{iminfo,Staffli-method}
\alias{iminfo,Staffli,Staffli-method}
\title{Extracts a list of image info from a Staffli object}
\usage{
iminfo(object)

iminfo(object)

\S4method{iminfo}{Staffli}(object)
}
\arguments{
\item{object}{A Staffli object}
}
\description{
Extracts a list of image info from a Staffli object

Method for extracting the HE image dimensions from a 'Staffli' object
}
