% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_from_table.R
\name{InputFromTable}
\alias{InputFromTable}
\title{Create S4 object from info table.
This function is a wrapper to create a complete S4 object with all the samples and metadata}
\usage{
InputFromTable(infotable, transpose = FALSE, topN = 0,
  min.gene.count = 0, min.gene.spots = 0, min.spot.feature.count = 0,
  min.spot.count = 0, annotation = NULL, id.column = NULL,
  replace.column = NULL, platform = "Visium",
  scaleVisium = 0.1039393, verbose = TRUE, ...)
}
\arguments{
\item{infotable}{table with paths to count files and metadata}

\item{topN}{OPTIONAL: Filter out the top most expressed genes}

\item{min.gene.count}{filter away genes that has a total read count below this threshold}

\item{min.gene.spots}{filter away genes that is not expressed below this number of capture spots}

\item{min.spot.feature.count}{filter away capture spots that contains a total feature count below this threshold}

\item{min.spot.count}{filter away capture spots that contains a total read count below this threshold}

\item{annotation}{data.frame with one id.column containing the current gene ids and a matching replace
column containing the new gene ids}

\item{id.column}{character vector of unique gene ids of the same type as in the Seurat object}

\item{replace.column}{character vector of unique gene ids to replace the old ids}

\item{scaleVisium}{10X visium scale factor for pixel coordinates, default set to 0.1039393}

\item{verbose}{Print messages}

\item{...}{parameters passed to \code{\link{CreateSeuratObejct}}}
}
\description{
Create S4 object from info table.
This function is a wrapper to create a complete S4 object with all the samples and metadata
}
