% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{labelRGB}
\alias{labelRGB}
\title{Select spots by rgb values after ST.DimPlot}
\usage{
labelRGB(object, dimPlot, label1 = "label1", label2 = NULL,
  red1 = c(0, 255), green1 = c(0, 255), blue1 = c(0, 255),
  red2 = c(0, 255), green2 = c(0, 255), blue2 = c(0, 255))
}
\arguments{
\item{object}{S4 object}

\item{dimPlot}{A ST.DimPlot output object}

\item{label1}{Optional name of the 1st group of capture-spots}

\item{label2}{Optional name of the 2nd group of capture-spots}
}
\description{
Select spots by rgb values after ST.DimPlot
}
