% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/image_processing.R
\name{ManualAlignImages}
\alias{ManualAlignImages}
\alias{ManualAlignImages.Staffli}
\alias{ManualAlignImages.Seurat}
\title{Manual alignment of images}
\usage{
ManualAlignImages(
  object,
  type = NULL,
  reference.index = 1,
  edges = TRUE,
  verbose = FALSE,
  maxnum = 1000,
  fix.axes = FALSE,
  custom.edge.detector = NULL
)

\method{ManualAlignImages}{Staffli}(
  object,
  type = "masked.masks",
  reference.index = 1,
  edges = TRUE,
  verbose = FALSE,
  maxnum = 1000,
  fix.axes = FALSE,
  custom.edge.detector = NULL
)

\method{ManualAlignImages}{Seurat}(
  object,
  type = "masked.masks",
  reference.index = 1,
  edges = TRUE,
  verbose = FALSE,
  maxnum = 1000,
  fix.axes = FALSE,
  custom.edge.detector = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{type}{Image type to use as input for alignment [default: 'masked.masks']}

\item{reference.index}{Specifies reference sample image for alignment [default: 1]}

\item{edges}{Uses the tissue edges for alignment}

\item{verbose}{Print messages}

\item{maxnum}{Maximum number of points to display in the app}

\item{custom.edge.detector}{Custom function used to detect edges in tissue image. If a function is provided, the
edges option will be overridden.}
}
\value{
A Staffli object

A Seurat object
}
\description{
Creates an interactive shiny application to align images manually
}
\examples{
# Create a new Staffli object, mask, align and plot images (will start an interactive shiny session)
st.obj <- CreateStaffliObject(imgs, meta.data)
st.obj <- LoadImages(st.obj, verbose = TRUE) \%>\% MaskImages() \%>\% ManualAlignImages()
plot(st.obj)
# Load, mask, align and plot images (will start an interactive shiny session)
se <- LoadImages(se, verbose = TRUE) \%>\% MaskImages() \%>\% ManualAlignImages()
ImagePlot(se)
}
