% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/image_processing.R
\name{MaskImages}
\alias{MaskImages}
\alias{MaskImages.Staffli}
\alias{MaskImages.Seurat}
\title{Masks the background of HE images stored in a Seurat object}
\usage{
MaskImages(object, iso.blur = 2, channels.use = 1, compactness = 1,
  add.contrast = TRUE, verbose = FALSE)

\method{MaskImages}{Staffli}(object, iso.blur = 2, channels.use = NULL,
  compactness = 1, add.contrast = NULL, verbose = FALSE)

\method{MaskImages}{Seurat}(object, iso.blur = 2, channels.use = NULL,
  compactness = 1, add.contrast = NULL, verbose = FALSE)
}
\arguments{
\item{object}{Seurat or Staffli object}

\item{iso.blur}{Sigma value (pixels) for isoblurring of HE images prior to image segmentation}

\item{channels.use}{Select channel to use for masking (default: 1)}

\item{compactness}{Controls scaling ratio for pixel values}

\item{verbose}{Print messages}
}
\value{
A Seurat object with masked HE images

A Staffli object

A Seurat object
}
\description{
Algorithm:
1. Blur image with median filter
2. Equalize image colors using an empirical cumulative distribution function
3. Convert image into SLIC superpixels using the \code{\link{slic}} function
4. Convert image to CIELAB colorspace and perform kmeans clustering (k = 2; inside/outside tissue) to segment image
5. Split segmented image into objects and filter out objects with a small area
6. Keep objects which overlaps with adjusted pixel coordinates
}
\examples{
# Create a new Staffli object, mask and plot images
st.obj <- CreateStaffliObject(imgs, meta.data)
st.obj <- LoadImages(st.obj, verbose = TRUE) \%>\% MaskImages()
plot(st.obj)


# Load images into a Seurat objectm, mask and plot images
se <- LoadImages(se, verbose = TRUE) \%>\% MaskImages()
ImagePlot(se)
}
