% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/image_processing.R
\name{MaskImages}
\alias{MaskImages}
\alias{MaskImages.Staffli}
\alias{MaskImages.Seurat}
\title{Image masking}
\usage{
MaskImages(object, iso.blur = 2, channels.use = 1, compactness = 1,
  add.contrast = TRUE, verbose = FALSE)

\method{MaskImages}{Staffli}(object, iso.blur = 2, channels.use = NULL,
  compactness = 1, add.contrast = NULL, verbose = FALSE,
  custom.msk.fkn = NULL)

\method{MaskImages}{Seurat}(object, iso.blur = 2, channels.use = NULL,
  compactness = 1, add.contrast = NULL, verbose = FALSE)
}
\arguments{
\item{object}{Seurat or Staffli object}

\item{iso.blur}{Sigma value (pixels) for isoblurring of HE images prior to image segmentation}

\item{channels.use}{Select channel to use for masking (default: 1)}

\item{compactness}{Scales the number of super-pixels}

\item{verbose}{Print messages}

\item{custom.msk.fkn}{Custom masking function that takes an image of class "cimg" as input and returns a mask
of class "pixset" outlining the tissue area.}
}
\value{
A Seurat object with masked HE images

A Staffli object

A Seurat object
}
\description{
Masks the background of a tissue section in HE images
}
\details{
The masking method provided uses the SLIC method; Simple Linear Iterative Clustering (for more info, see
\url{https://www.r-bloggers.com/superpixels-in-imager/}). This algorithm generates superpixels by
clustering pixels based on their color similarity and proximity in the image plane.
Before running the SLIC method on the HE images we have noticed that some pre-processing can improve the
masking significantly.
First of all, it can be beneficial to remove 1 or two color channels from the image
using \code{channels.use}. Specifying \code{channels.use = 1} will keep only the first channel before running
SLIC.
The next step is to apply some blurring efect
}
\section{Masking options}{

}

\section{Custom mask function}{

Masking can sometimes very difficult to accomplish without cutting off parts of the tissue or
including unwanted parts of the tissue. If the masking fails you also have the option to
write your own function that takes a "cimg" class image as input and returns a masked "pxset".
}

\examples{
# Create a new Staffli object, mask and plot images
st.obj <- CreateStaffliObject(imgs, meta.data)
st.obj <- LoadImages(st.obj, verbose = TRUE) \%>\% MaskImages()
plot(st.obj)


# Load images into a Seurat objectm, mask and plot images
se <- LoadImages(se, verbose = TRUE) \%>\% MaskImages()
ImagePlot(se)
}
