% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{obtain.array.coords}
\alias{obtain.array.coords}
\title{Obtain array coordinates}
\usage{
obtain.array.coords(st.object, data, image.type)
}
\arguments{
\item{object}{Seurat object}
}
\description{
Obtain array coordinates
}
