% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STUtility.R, R/hello.R
\name{palette.select}
\alias{palette.select}
\title{Palette selection}
\usage{
palette.select(palette)

palette.select(palette)
}
\arguments{
\item{palette}{Palette choice for plotting spatial expression histology heatmap}

\item{palette}{Palette choice for plotting spatial expression histology heatmap}
}
\description{
Palette selection

Palette selection
}
\keyword{internal}
