% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{palette.select}
\alias{palette.select}
\title{Palette selection}
\usage{
palette.select(palette, info = F)
}
\arguments{
\item{palette}{Palette choice for plotting spatial expression histology heatmap}
}
\description{
Palette selection
}
\keyword{internal}
