% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_from_table.R
\name{parse.spot.file}
\alias{parse.spot.file}
\title{load count files}
\usage{
parse.spot.file(path, delim = "\\t", ...)
}
\arguments{
\item{path}{Path to count files}

\item{delim}{delimiter}

\item{suffix}{add suffix to path}

\item{row.names}{set column as row.names}

\item{path}{Path to spot files}
}
\description{
load count files
}
\keyword{Parse}
\keyword{ST}
\keyword{detector}
\keyword{from}
\keyword{internal}
\keyword{output}
\keyword{spot}
\keyword{the}
\keyword{tool}
