% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEH.R, R/STUtility.R
\name{plotJ}
\alias{plotJ}
\title{======================================= INTERNAL ====================================
 plot tile list for voronoi}
\usage{
plotJ(x, verbose = FALSE, close = FALSE, pch = 1,
  fillcol = getCol(x, warn = warn), col.pts = NULL, col.num = NULL,
  border = NULL, showpoints = !number, add = FALSE, asp = 1,
  clipp = NULL, xlab = "x", ylab = "y", main = "", warn = FALSE,
  number = FALSE, adj = NULL, axes = FALSE, ...)

plotJ(x, verbose = FALSE, close = FALSE, pch = 1,
  fillcol = getCol(x, warn = warn), col.pts = NULL, col.num = NULL,
  border = NULL, showpoints = !number, add = FALSE, asp = 1,
  clipp = NULL, xlab = "x", ylab = "y", main = "", warn = FALSE,
  number = FALSE, adj = NULL, axes = FALSE, ...)
}
\description{
======================================= INTERNAL ====================================
 plot tile list for voronoi

plot tile list for voronoi
}
\keyword{internal}
