% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_from_table.R
\name{prep.from.table}
\alias{prep.from.table}
\title{Create S4 object from info table.
This function is a wrapper to create a complete S4 object with all the samples and metadata}
\usage{
prep.from.table(infotable, transpose = TRUE, topN = 0,
  min.gene.count = 0, min.gene.spots = 0, min.spot.count = 0,
  spot.file = TRUE, annotation = NULL, id.column = NULL,
  replace.column = NULL, ...)
}
\arguments{
\item{infotable}{table with paths to count files and metadata}

\item{transpose}{set TRUE if count files are in the form of genes as columns and spatial
coordiantes as rows.}

\item{topN}{OPTIONAL: Filter out the top most expressed genes}

\item{min.gene.count}{filter away genes that has a total read count below this threshold}

\item{min.gene.spots}{filter away genes that is not expressed below this number of capture spots}

\item{min.spot.count}{filter away capture spots that contains a total read count below this threshold}

\item{annotation}{data.frame with one id.column containing the current gene ids and a matching replace
column containing the new gene ids}

\item{id.column}{character vector of unique gene ids of the same type as in the Seurat object}

\item{replace.column}{character vector of unique gene ids to replace the old ids}

\item{...}{parameters passed to \code{\link{CreateSeuratObejct}}}
}
\description{
Create S4 object from info table.
This function is a wrapper to create a complete S4 object with all the samples and metadata
}
