% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STUtility.R
\name{prepare.counts}
\alias{prepare.counts}
\title{Create S4 object}
\usage{
prepare.counts(sampleTable, transpose = TRUE, topN = 0, th.gene = 0,
  th.spot = 0, object.type = "Seurat", ...)
}
\arguments{
\item{sampleTable}{table with paths to count files and metadata}

\item{transpose}{set TRUE if count files are in the form of genes as columns and spatial
coordiantes as rows.}

\item{topN}{OPTIONAL: Filter out the top most expressed genes}

\item{th.gene}{OPTIONAL: Filter genes with total counts across all samples below threshold}

\item{th.spot}{OPTIONAL: Filter `"capture-spots"` with total count values below threshold}
}
\description{
Create S4 object
}
