% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{qcGenes}
\alias{qcGenes}
\title{Quality metrics for genes}
\usage{
qcGenes(exprMat, delim = "x", labels = NULL)
}
\arguments{
\item{exprMat}{expression matrix with spots in columns and genes in rows}

\item{delim}{delimiter}

\item{labels}{character vector with labels corresponding to datasets in merged expression matrix}
}
\value{
data.frame with gene names and quality metrics
}
\description{
Quality metrics for genes
}
