% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{rasterlists}
\alias{rasterlists}
\title{Extracts available image types from Seurat or Staffli objects}
\usage{
rasterlists(object)
}
\arguments{
\item{object}{A Seurat or Staffli object}
}
\description{
Extracts available image types from Seurat or Staffli objects
}
