% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rigid_transformation.R
\name{rigid.rot}
\alias{rigid.rot}
\title{Creates a transformation matrix for rotation}
\usage{
rigid.rot(alpha = 0, forward = TRUE)
}
\arguments{
\item{alpha}{rotation angle}
}
\description{
Creates a transformation matrix for clockwise rotation by 'alpha' degrees
}
