% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunNMF.R
\name{RunNMF}
\alias{RunNMF}
\title{Run Non-negative Matrix Factorization}
\usage{
RunNMF(
  object,
  assay = NULL,
  slot = "scale.data",
  features = NULL,
  nfactors = 20,
  rescale = TRUE,
  reduction.name = "NMF",
  reduction.key = "factor_",
  n.cores = NULL,
  order.by.spcor = FALSE,
  sort.spcor.by.var = FALSE,
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{assay}{Assay Name of Assay NMF is being run on}

\item{slot}{Slot to pull data from.}

\item{features}{Features to compute the NMF for. Note that these features must be present in the
slot used to compute the NMF. By default, the `features` is set to `VariableFeatures(object)`
to include the most variable features selected in the normalization step.}

\item{nfactors}{Total Number of factors to compute and store (20 by default)}

\item{rescale}{Rescale data to make sure that values of the input matrix are non-n}

\item{reduction.name}{Dimensional reduction name, "NMF" by default}

\item{reduction.key}{Dimensional reduction key, specifies the prefix of the factor ids, e.g.
"factor_1", "factor_2", etc.}

\item{n.cores}{Number of threads to use in computation}

\item{order.by.spcor}{Order factors by spatial correlation}

\item{sort.spcor.by.var}{Sort factors by decreasing variance}

\item{...}{Additional parameters}
}
\description{
Decompose an expression matrix A with non-negative elements into matrices WxH, also with
non-negative elements. W is the feature loading matrix (features x factors) and H is the
low dimensional embedding of the spots (factors x spots).
}
