% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{samplenames}
\alias{samplenames}
\title{Extracts available samplenames from Seurat or Staffli objects}
\usage{
samplenames(object)
}
\arguments{
\item{object}{A Seurat or Staffli object}
}
\description{
Extracts available samplenames from Seurat or Staffli objects
}
