% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STUtility.R, R/hello.R
\name{save.table}
\alias{save.table}
\title{Gene cluster tables}
\usage{
save.table(se, filename)

save.table(se, filename)
}
\description{
Creates tables of the gene clusters, both with raw count data,
and the relative freq. data used in the heatmap plots.

Creates tables of the gene clusters, both with raw count data,
and the relative freq. data used in the heatmap plots.
}
