% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEH.R
\name{save.table}
\alias{save.table}
\title{Gene cluster tables}
\usage{
save.table(se, filename)
}
\arguments{
\item{se}{s4 object}

\item{filename}{filename}
}
\description{
Creates tables of the gene clusters, both with raw count data,
and the relative freq. data used in the heatmap plots.
}
