% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/Staffli.R
\docType{methods}
\name{scaled.imdims}
\alias{scaled.imdims}
\alias{scaled.imdims,Staffli-method}
\alias{scaled.imdims,Staffli,Staffli-method}
\title{Extracts scaled dimensions of images in Staffli object}
\usage{
scaled.imdims(object, type = "raw")

scaled.imdims(object, type = "raw")

\S4method{scaled.imdims}{Staffli}(object, type = "raw")
}
\arguments{
\item{object}{A Staffli object}

\item{type}{Image type [choices: 'raw', 'masked', 'processed']}
}
\description{
Extracts scaled dimensions of images in Staffli object

Method for extracting the scaled HE image dimensions from a 'Staffli' object
}
