% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STUtility.R, R/hello.R
\name{seh.tsne}
\alias{seh.tsne}
\title{Create tSNE based on spatial expression histology}
\usage{
seh.tsne(se, filename, perplexity = 30, priorPCA = FALSE,
  bg = "black")

seh.tsne(se, filename, perplexity = 30, priorPCA = FALSE,
  bg = "black")
}
\description{
Create tSNE based on spatial expression histology

Create tSNE based on spatial expression histology
}
