% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{ST.DimPlot}
\alias{ST.DimPlot}
\title{Dimensional reduction plot on ST coordinates}
\usage{
ST.DimPlot(object, dims = c(1, 2), spots = NULL, blend = FALSE,
  min.cutoff = NA, max.cutoff = NA, pt.size = 1, reduction = NULL,
  group.by = NULL, shape.by = NULL, palette = "MaYl", rev.cols = F,
  dark.theme = F, ncol = NULL, delim = NULL, return.plot.list = F,
  grid.ncol = NULL, center.zero = T, channels.use = NULL, ...)
}
\arguments{
\item{object}{Seurat object}

\item{dims}{Dimensions to plot, must numeric vectoir specifying number of dimensions to plot}

\item{spots}{Vector of spots to plot (default is all spots)}

\item{blend}{Scale and blend expression values to visualize coexpression of two features (This options will override other coloring parameters)}

\item{min.cutoff, max.cutoff}{Vector of minimum and maximum cutoff values for each feature,
may specify quantile in the form of 'q##' where '##' is the quantile (eg, 'q1', 'q10')}

\item{pt.size}{Adjust point size for plotting}

\item{reduction}{Which dimensionality reduction to use. If not specified, first searches for umap, then tsne, then pca}

\item{group.by}{Name of a metadata column to facet plot by (deault is sampleID)}

\item{shape.by}{If NULL, all points are circles (default). You can specify any
cell attribute (that can be pulled with FetchData) allowing for both
different colors and different shapes on cells}

\item{palette}{color palette used for spatial heatmap}

\item{rev.cols}{logical specifying whether colorscale should be reversed}

\item{ncol}{number of columns in \code{facet_wrap}}

\item{delim}{delimiter passed to \code{\link{GetCoords}} if adjusted ST coordinates are missing in the meta data}

\item{return.plot.list}{should the plots be returned as a list? By default, the plots are arranged into a grid}

\item{grid.ncol}{Number of columns for display when combining plots}

\item{center.zero}{should the colorscale be centered around 0? Set to TRUE for scaled data}

\item{...}{Extra parameters passed on to \code{\link{STPlot}}}
}
\value{
A ggplot object
}
\description{
Graphs the output of a dimensional reduction technique on a 2D scatter plot where each point is a
spot and it's positioned based on the ST array coordinates in a grid pattern. Spots are colored
by the dimensional reduction vectors specified by the user.
}
\details{
Function built upon the DimPlot() function from Seurat (https://github.com/satijalab/seurat/blob/master/R/utilities.R)
}
