% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{ST.DimPlot}
\alias{ST.DimPlot}
\title{Dimensional reduction plot on ST coordinates}
\usage{
ST.DimPlot(object, dims = c(1, 2), spots = NULL, plot.type = "spots",
  blend = FALSE, min.cutoff = NA, max.cutoff = NA, pt.size = 1,
  pt.alpha = 1, reduction = NULL, shape.by = NULL,
  palette = "MaYl", cols = NULL, rev.cols = FALSE,
  dark.theme = FALSE, ncol = NULL, grid.ncol = NULL,
  center.zero = T, channels.use = NULL, center.tissue = FALSE,
  verbose = FALSE, ...)
}
\arguments{
\item{object}{Seurat object}

\item{dims}{Dimensions to plot, a numeric vector specifying number of dimensions to plot}

\item{spots}{Vector of spots to plot (default is all spots)}

\item{plot.type}{Specify the type of plot to use (default: "spots"). Available options are; "spots", "smooth" and "tesselation"}

\item{blend}{Scale and blend expression values to visualize coexpression of two features (This options will override other coloring parameters)}

\item{min.cutoff, max.cutoff}{Vector of minimum and maximum cutoff values for each feature,
may specify quantile in the form of 'q##' where '##' is the quantile (eg, 'q1', 'q10')}

\item{pt.size}{Adjust point size for plotting}

\item{pt.alpha}{Adjust point opacity for plotting}

\item{reduction}{Which dimensionality reduction to use. If not specified, first searches for umap, then tsne, then pca}

\item{shape.by}{If NULL, all points are circles (default). You can specify any
cell attribute (that can be pulled with FetchData) allowing for both
different colors and different shapes on cells}

\item{palette}{Color palette used for spatial heatmap (see \code{?palette.select} for available options).
Disabled if a color vector is provided (see \code{cols} below).}

\item{cols}{A vector of colors to use for colorscale, e.g. \code{cols = c("blue", "white", "red")} will
create a gradient color scale going from blue to white to red. This options will deactivate the \code{palette}
option.}

\item{rev.cols}{Logical specifying whether colorscale should be reversed}

\item{ncol}{Number of columns in \code{facet_wrap}}

\item{grid.ncol}{Number of columns for display when combining plots}

\item{center.zero}{Should the colorscale be centered around 0? Set to TRUE for scaled data}

\item{center.tissue}{Adjust coordinates so that the center of the tissue is in the middle of the array along the y-axis}

\item{verbose}{Print messages}

\item{...}{Extra parameters passed on to \code{\link{STPlot}}}
}
\value{
A ggplot object
}
\description{
Graphs the output of a dimensional reduction technique on a 2D scatter plot where each point is a
spot and it's positioned based on the ST array coordinates in a grid pattern. Spots are colored
by the dimensional reduction vectors specified by the user.
}
\details{
Function built upon the DimPlot() function from Seurat (https://github.com/satijalab/seurat/blob/master/R/utilities.R)
}
