% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{ST.FeaturePlot}
\alias{ST.FeaturePlot}
\title{Visualize 'features' on an ST array grid}
\usage{
ST.FeaturePlot(object, features, split.labels = FALSE, indices = NULL,
  spots = NULL, plot.type = "spots", min.cutoff = NA,
  max.cutoff = NA, slot = "data", blend = FALSE, pt.size = 1,
  shape.by = NULL, palette = NULL, rev.cols = FALSE,
  dark.theme = FALSE, ncol = NULL, delim = NULL,
  return.plot.list = FALSE, grid.ncol = NULL, center.zero = FALSE,
  channels.use = NULL, center.tissue = FALSE, theme = theme_void(),
  verbose = FALSE, ...)
}
\arguments{
\item{object}{Seurat object}

\item{features}{\itemize{
    \item An \code{Assay} feature (e.g. a gene name - "MS4A1")
    \item A column name from meta.data (e.g. mitochondrial percentage - "percent.mito")
}}

\item{split.labels}{Only active if the features are specified by character vectors.
The plot will be split into one plot for each group label, highlighting the labelled spots.}

\item{indices}{Numeric vector specifying sample indices to include in plot. Default is to show all samples.}

\item{spots}{Vector of spots to plot (default is all spots)}

\item{plot.type}{Specify the type of plot to use (default: "spots"). Available options are; "spots", "smooth" and "tesselation"}

\item{min.cutoff, max.cutoff}{Vector of minimum and maximum cutoff values for each feature,
may specify quantile in the form of 'q##' where '##' is the quantile (eg, 'q1', 'q10')}

\item{slot}{Which slot to pull expression data from?}

\item{blend}{Scale and blend expression values to visualize coexpression of two features (This options will override other coloring parameters)}

\item{pt.size}{Adjust point size for plotting}

\item{shape.by}{If NULL, all points are circles (default). You can specify any
cell attribute (that can be pulled with FetchData) allowing for both
different colors and different shapes on cells}

\item{palette}{color palette used for spatial heatmap}

\item{rev.cols}{logical specifying whether colorscale should be reversed}

\item{ncol}{number of columns in \code{facet_wrap}}

\item{delim}{delimiter passed to \code{\link{GetCoords}} if adjusted ST coordinates are missing in the meta data}

\item{return.plot.list}{should the plots be returned as a list? By default, the plots are arranged into a grid}

\item{grid.ncol}{Number of columns for display when combining plots}

\item{center.zero}{should the colorscale be centered around 0? Set to TRUE for scaled data}

\item{center.tissue}{Adjust coordinates so that the center of the tissue is in the middle of the array along the y-axis}

\item{theme}{Object of class "theme" used to change the background theme}

\item{verbose}{Print messages}

\item{...}{Extra parameters passed on to \code{\link{STPlot}}}
}
\value{
A ggplot object
}
\description{
Colors spots on an an ST array grid according to a 'feature'
(i.e. gene expression (raw counts or scaled) and features available in the meta data slot)
}
\details{
Function built upon the FeaturePlot() function from Seurat (https://github.com/satijalab/seurat/blob/master/R/visualization.R)
}
