% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{ST.ImagePlot}
\alias{ST.ImagePlot}
\title{Graphs ST spots colored by continuous variable, e.g. dimensional reduction vector}
\usage{
ST.ImagePlot(
  data,
  data.type,
  shape.by,
  variable,
  image,
  dims,
  pt.size = 2,
  pt.alpha = 1,
  pt.border = FALSE,
  add.alpha = FALSE,
  palette = "MaYl",
  cols = NULL,
  ncol = NULL,
  spot.colors = NULL,
  center.zero = TRUE,
  plot.title = NULL,
  dark.theme = FALSE,
  pixels.per.um = NULL,
  limits = NULL,
  custom.theme = NULL,
  ...
)
}
\arguments{
\item{data}{Object of class 'data.frame' containing at least (x, y) coordinates, a "sample" vector with labels for each sample
and one column with the feature values. Can also include an additional column for shapes.}

\item{data.type}{String specifying the class of the features in data to be plotted}

\item{shape.by}{String specifying the column where the shaping label is stored}

\item{variable}{Name of feature column}

\item{image}{Image of class "raster" to use as background for plotting}

\item{dims}{List of dimensions for original images. This list has to contain one element for each sample and each element
should be a vector of length 2 specifying the dimensions of the original HE image.}

\item{pt.size}{Point size of each ST spot [default: 1]}

\item{pt.alpha}{Opacity of each ST spot [default: 1]}

\item{pt.border}{Should a border be drawn around the spots? [default: TRUE]}

\item{palette}{Color palette used for spatial heatmap (see \code{palette.select(info = T)} for available options).
Disabled if a color vector is provided (see \code{cols} below).}

\item{cols}{A vector of colors to use for colorscale, e.g. \code{cols = c("blue", "white", "red")} will
create a gradient color scale going from blue to white to red. This options will deactivate the \code{palette}
option.}

\item{ncol}{Number of columns to arrange the samples into. This can for example be useful to adjust if you want to visualize the samples
in just in one row or one column.}

\item{spot.colors}{Character vector with color names that overrides default coloring with ggplot2}

\item{center.zero}{Specifies whther or not the colorscale should be centered around 0. For some values, such as Principal Component vectors,
the distribution of values is centered at 0 and in that case it can be appropriate to use a divergent colorscale with a predefined value for 0.
If this parameter is set to TRUE, the ggplot2 function \code{scale_color_gradient2} will be used to control the coloring instead of
\code{scale_color_gradientn}. If center.zero is set to FALSE, the colorscale will simply map the values in equally spaced intervals which could skew
the interpretaion of the output plot.}

\item{plot.title}{String specifying the title of the plot}

\item{dark.theme}{Switches color of scalebar to 'white'}

\item{pixels.per.um}{Defines the number of pixels per micrometer to draw the scale bar}

\item{limits}{Sets the limits of the colorbar}

\item{custom.theme}{Object of class 'theme' used to change the background theme (see for example \url{https://ggplot2.tidyverse.org/reference/theme.html})}

\item{...}{Parameters passed to geom_point()}
}
\description{
Graphs ST spots colored by continuous variable, e.g. dimensional reduction vector
}
