% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STUtility.R, R/hello.R
\name{st.load.matrix}
\alias{st.load.matrix}
\title{load count files}
\usage{
st.load.matrix(path, suffix = "", row.names = 1, ...)

st.load.matrix(path, suffix = "", row.names = 1, ...)
}
\arguments{
\item{path}{Path to count files}

\item{suffix}{add suffix to path}

\item{row.names}{set column as row.names}

\item{path}{Path to count file}

\item{suffix}{add suffix to path}

\item{row.names}{set column as row.names}
}
\description{
load count files

load count files
}
\keyword{internal}
