% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{STPlot}
\alias{STPlot}
\title{Graphs ST spots colored by continuous variable, e.g. dimensional reduction vector}
\usage{
STPlot(data, data.type, group.by, shape.by, variable, pt.size = 1,
  palette = "MaYl", rev.cols = F, ncol = NULL, spot.colors = NULL,
  center.zero = T, plot.title = NULL, ...)
}
\arguments{
\item{data}{data.frame containing (x, y) coordinates, a group vector and a continuous variable vector}

\item{data.type}{type of data, e.g. numeric or integer}

\item{group.by}{specifies column to facet the plots by, e.g. sampleID}

\item{shape.by}{specifies column to shape points by, e.g. morphological region}

\item{variable}{name of continuous variable}

\item{pt.size}{point size of each ST spot}

\item{palette}{color palette used for spatial heatmap}

\item{rev.cols}{logical specifying whether colorscale should be reversed}

\item{ncol}{number of columns in \code{facet_wrap}}

\item{spot.colors}{character vector woth color names that overrides default coloring with ggplot2}

\item{center.zero}{should the colorscale be centered around 0? Set to TRUE for scaled data}

\item{...}{parameters passed to geom_point()}
}
\description{
Graphs ST spots colored by continuous variable, e.g. dimensional reduction vector
}
