% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{STPlot}
\alias{STPlot}
\title{Graphs ST spots colored by continuous variable, e.g. dimensional reduction vector}
\usage{
STPlot(data, data.type = NULL, group.by, shape.by, variable,
  pt.size = 1, palette = "MaYl", rev.cols = F, ncol = NULL,
  spot.colors = NULL, center.zero = TRUE, center.tissue = F,
  plot.title = NULL, xlim = NULL, ylim = NULL,
  split.labels = FALSE, theme = theme_void(), ...)
}
\arguments{
\item{data}{data.frame containing (x, y) coordinates, a group vector and a continuous variable vector}

\item{data.type}{type of data, e.g. numeric or integer}

\item{group.by}{specifies column to facet the plots by, e.g. sample}

\item{shape.by}{specifies column to shape points by, e.g. morphological region}

\item{variable}{name of continuous variable}

\item{pt.size}{point size of each ST spot}

\item{palette}{color palette used for spatial heatmap}

\item{rev.cols}{logical specifying whether colorscale should be reversed}

\item{ncol}{number of columns in \code{facet_wrap}}

\item{spot.colors}{character vector woth color names that overrides default coloring with ggplot2}

\item{center.zero}{should the colorscale be centered around 0? Set to TRUE for scaled data}

\item{center.tissue}{Adjust coordinates so that the center of the tissue is in the middle of the array along the y-axis}

\item{plot.title}{Add title to plot}

\item{xlim, ylim}{Set x/y-axis limits}

\item{split.labels}{Only active if the features are specified by character vectors.
The plot will be split into one plot for each group label, highlighting the labelled spots.}

\item{theme}{Object of class "theme" used to change the background theme}

\item{...}{parameters passed to geom_point()}
}
\description{
Graphs ST spots colored by continuous variable, e.g. dimensional reduction vector
}
