% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{STPlot}
\alias{STPlot}
\title{Graphs ST spots colored by continuous variable, e.g. dimensional reduction vector}
\usage{
STPlot(data, data.type = NULL, group.by, shape.by, variable,
  pt.size = 1, palette = "MaYl", cols = NULL, rev.cols = F,
  ncol = NULL, spot.colors = NULL, center.zero = TRUE,
  center.tissue = F, plot.title = NULL, dims = NULL,
  split.labels = FALSE, theme = theme_void(), ...)
}
\arguments{
\item{data}{data.frame containing (x, y) coordinates, a group vector and a continuous variable vector}

\item{data.type}{Type of data, e.g. numeric or integer}

\item{group.by}{Specifies column to facet the plots by, e.g. sample}

\item{shape.by}{Specifies column to shape points by, e.g. morphological region}

\item{variable}{Name of continuous variable}

\item{pt.size}{Point size of each ST spot}

\item{palette}{Color palette used for spatial heatmap (see \code{?palette.select} for available options).
Disabled if a color vector is provided (see \code{cols} below).}

\item{cols}{A vector of colors to use for colorscale, e.g. \code{cols = c("blue", "white", "red")} will
create a gradient color scale going from blue to white to red. This options will deactivate the \code{palette}
option.}

\item{rev.cols}{Logical specifying whether colorscale should be reversed}

\item{ncol}{Number of columns in \code{facet_wrap}}

\item{spot.colors}{Character vector woth color names that overrides default coloring with ggplot2}

\item{center.zero}{Should the colorscale be centered around 0? Set to TRUE for scaled data}

\item{center.tissue}{Adjust coordinates so that the center of the tissue is in the middle of the array along the y-axis}

\item{plot.title}{Add title to plot}

\item{dims}{List of dimensions for x and y scales}

\item{split.labels}{Only active if the features are specified by character vectors.
The plot will be split into one plot for each group label, highlighting the labelled spots.}

\item{theme}{Object of class "theme" used to change the background theme}

\item{...}{Parameters passed to geom_point()}
}
\description{
Graphs ST spots colored by continuous variable, e.g. dimensional reduction vector
}
