% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R, R/manual_annotation.R,
%   R/pre_processing.R, R/process_from_table.R
\docType{package}
\name{STutility-package}
\alias{STutility}
\alias{STutility-package}
\title{STutility: Data analysis and image handing of Spatial Transcriptomics data}
\description{
STutility collects usful tools for Spatial Transcriptomics data analysis workflow within R.
    The package depends on the Seurat R package, which is a toolkit for analysis and exploration of single cell RNA sequencing data (scRNAseq).
	STutility adds image functionallity to the Seurat workflow, allowing RNAseq data to be put in a spatial tissue context.
	Best practices for data transformation, QC and typical analysis workflows are demonstrated.

STutility collects usful tools for Spatial Transcriptomics data analysis workflow within R.
    The package depends on the Seurat R package, which is a toolkit for analysis and exploration of single cell RNA sequencing data (scRNAseq).
	STutility adds image functionallity to the Seurat workflow, allowing RNAseq data to be put in a spatial tissue context.
	Best practices for data transformation, QC and typical analysis workflows are demonstrated.

STutility collects usful tools for Spatial Transcriptomics data analysis workflow within R.
    The package depends on the Seurat R package, which is a toolkit for analysis and exploration of single cell RNA sequencing data (scRNAseq).
	STutility adds image functionallity to the Seurat workflow, allowing RNAseq data to be put in a spatial tissue context.
	Best practices for data transformation, QC and typical analysis workflows are demonstrated.

STutility collects usful tools for Spatial Transcriptomics data analysis workflow within R.
    The package depends on the Seurat R package, which is a toolkit for analysis and exploration of single cell RNA sequencing data (scRNAseq).
	STutility adds image functionallity to the Seurat workflow, allowing RNAseq data to be put in a spatial tissue context.
	Best practices for data transformation, QC and typical analysis workflows are demonstrated.
}
\author{
\strong{Maintainer}: Joseph Bergenstraahle \email{j.bergenstrahle@scilifelab.se}

Authors:
\itemize{
  \item Ludvig Larsson \email{ludvig.larsson@scilifelab.se}
}


\strong{Maintainer}: Joseph Bergenstraahle \email{j.bergenstrahle@scilifelab.se}

Authors:
\itemize{
  \item Ludvig Larsson \email{ludvig.larsson@scilifelab.se}
}


\strong{Maintainer}: Joseph Bergenstraahle \email{j.bergenstrahle@scilifelab.se}

Authors:
\itemize{
  \item Ludvig Larsson \email{ludvig.larsson@scilifelab.se}
}


\strong{Maintainer}: Joseph Bergenstraahle \email{j.bergenstrahle@scilifelab.se}

Authors:
\itemize{
  \item Ludvig Larsson \email{ludvig.larsson@scilifelab.se}
}

}
