% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Staffli.R
\name{SubsetSTData}
\alias{SubsetSTData}
\title{Subset a Seurat object containing Staffli image data}
\usage{
SubsetSTData(
  object,
  expression,
  spots = NULL,
  features = NULL,
  idents = NULL,
  ...
)
}
\arguments{
\item{object}{A Seurat object containing Staffli data}

\item{expression}{Logical expression indicating features/variables to keep}

\item{spots}{A vector of spots to keep}

\item{features}{A vector of features to keep}

\item{idents}{A vector of identity classes to keep}

\item{...}{Extra parameters passed to WhichCells, such as slot, invert, or downsample}
}
\description{
Subsets a Seurat object containing Spatial Transcriptomics data while
making sure that the images and the spot coordinates are subsetted correctly.
If you use the default \code{\link{subset}} function there is a risk that images
are kept in the output Seurat object which will make the STUtility functions
crash.
}
\examples{
# Subset using meta data to keep spots with more than 1000 unique genes
se.subset <- SubsetSTData(se, expression = nFeature_RNA >= 1000)

# Subset by a predefined set of spots 
se.subset <- SubsetSTData(se, spots = keep.spots)

# Subset by a predefined set of features
se.subset <- SubsetSTData(se, features = keep.features)
}
